Refactor main.go, nicer help, feedback on config updates.
