+++ cmd/earthly/main.go
@@ -1,83 +1,47 @@
 package main
 
 import (
-	"bufio"
 	"bytes"
 	"context"
 	"crypto/rand"
 	"encoding/base64"
-	"encoding/json"
 	"flag"
 	"fmt"
 	"io"
-	"net/http"
 	_ "net/http/pprof" // enable pprof handlers on net/http listener
 	"net/url"
 	"os"
 	"os/signal"
-	"os/user"
 	"path"
 	"path/filepath"
 	"regexp"
 	"runtime"
-	"runtime/debug"
-	"sort"
 	"strconv"
 	"strings"
 	"syscall"
-	"text/tabwriter"
 	"time"
 
-	"github.com/containerd/containerd/platforms"
-	"github.com/dustin/go-humanize"
 	gsysinfo "github.com/elastic/go-sysinfo"
 	"github.com/fatih/color"
 	"github.com/joho/godotenv"
-	"github.com/moby/buildkit/client"
 	_ "github.com/moby/buildkit/client/connhelper/dockercontainer" // Load "docker-container://" helper.
-	"github.com/moby/buildkit/client/llb"
-	gwclient "github.com/moby/buildkit/frontend/gateway/client"
-	"github.com/moby/buildkit/session"
-	"github.com/moby/buildkit/session/auth/authprovider"
-	"github.com/moby/buildkit/session/localhost/localhostprovider"
-	"github.com/moby/buildkit/session/sshforward/sshprovider"
-	"github.com/moby/buildkit/util/entitlements"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
 	"github.com/urfave/cli/v2"
-	"golang.org/x/crypto/ssh"
-	"golang.org/x/crypto/ssh/agent"
-	"golang.org/x/sync/errgroup"
-	"golang.org/x/term"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 
 	"github.com/earthly/earthly/analytics"
-	"github.com/earthly/earthly/ast"
-	"github.com/earthly/earthly/autocomplete"
-	"github.com/earthly/earthly/buildcontext"
-	"github.com/earthly/earthly/buildcontext/provider"
 	"github.com/earthly/earthly/builder"
 	"github.com/earthly/earthly/buildkitd"
-	"github.com/earthly/earthly/cleanup"
 	"github.com/earthly/earthly/cloud"
 	"github.com/earthly/earthly/config"
 	"github.com/earthly/earthly/conslogging"
-	debuggercommon "github.com/earthly/earthly/debugger/common"
-	"github.com/earthly/earthly/debugger/terminal"
-	"github.com/earthly/earthly/docker2earthly"
-	"github.com/earthly/earthly/domain"
 	"github.com/earthly/earthly/earthfile2llb"
-	"github.com/earthly/earthly/states"
 	"github.com/earthly/earthly/util/cliutil"
 	"github.com/earthly/earthly/util/containerutil"
 	"github.com/earthly/earthly/util/fileutil"
-	"github.com/earthly/earthly/util/llbutil/secretprovider"
-	"github.com/earthly/earthly/util/platutil"
 	"github.com/earthly/earthly/util/reflectutil"
-	"github.com/earthly/earthly/util/syncutil/semutil"
-	"github.com/earthly/earthly/util/termutil"
-	"github.com/earthly/earthly/variables"
 )
 
 const (
@@ -194,17 +158,6 @@ var (
 	// BuiltBy contains information on which build-system was used (e.g. official earthly binaries, homebrew, etc)
 	BuiltBy string
 )
-
-var (
-	errLoginFlagsHaveNoEffect            = errors.New("account login flags have no effect when --auth-token (or the EARTHLY_TOKEN environment variable) is set")
-	errLogoutHasNoEffectWhenAuthTokenSet = errors.New("account logout has no effect when --auth-token (or the EARTHLY_TOKEN environment variable) is set")
-)
-
-func profhandler() {
-	addr := "127.0.0.1:6060"
-	fmt.Printf("listening for pprof on %s\n", addr)
-	http.ListenAndServe(addr, nil)
-}
 
 func main() {
 	startTime := time.Now()
@@ -377,7 +330,7 @@ func newEarthlyApp(ctx context.Context, console conslogging.ConsoleLogger) *eart
 
 	earthly := getBinaryName()
 
-	app.cliApp.Usage = "A build automation tool for the container era"
+	app.cliApp.Usage = "The CI/CD framework that runs anywhere"
 	app.cliApp.UsageText = "\t" + earthly + " [options] <target-ref>\n" +
 		"\n" +
 		"   \t" + earthly + " [options] --image <target-ref>\n" +
@@ -386,804 +339,23 @@ func newEarthlyApp(ctx context.Context, console conslogging.ConsoleLogger) *eart
 		"\n" +
 		"   \t" + earthly + " [options] command [command options]\n" +
 		"\n" +
-		"Executes Earthly builds. For more information see https://docs.earthly.dev/earthly-command.\n" +
-		"To get started with using Earthly, check out the getting started guide at https://docs.earthly.dev/guides/basics."
+		"Executes Earthly builds. For more information see https://docs.earthly.dev/docs/earthly-command.\n" +
+		"To get started with using Earthly, check out the getting started guide at https://docs.earthly.dev/basics.\n" +
+		"\n" +
+		"For help on build-specific flags try \n" +
+		"\n" +
+		"\t" + earthly + " build --help"
 	app.cliApp.UseShortOptionHandling = true
 	app.cliApp.Action = app.actionBuild
 	app.cliApp.Version = getVersionPlatform()
-	app.cliApp.Flags = []cli.Flag{
-		&cli.StringSliceFlag{
-			Name:    "platform",
-			EnvVars: []string{"EARTHLY_PLATFORMS"},
-			Usage:   "Specify the target platform to build for",
-			Value:   &app.platformsStr,
-		},
-		&cli.StringSliceFlag{
-			Name:    "build-arg",
-			EnvVars: []string{"EARTHLY_BUILD_ARGS"},
-			Usage:   "A build arg override, specified as <key>=[<value>]",
-			Value:   &app.buildArgs,
-		},
-		&cli.StringSliceFlag{
-			Name:    "secret",
-			Aliases: []string{"s"},
-			EnvVars: []string{"EARTHLY_SECRETS"},
-			Usage:   "A secret override, specified as <key>=[<value>]",
-			Value:   &app.secrets,
-		},
-		&cli.StringSliceFlag{
-			Name:    "secret-file",
-			EnvVars: []string{"EARTHLY_SECRET_FILES"},
-			Usage:   "A secret override, specified as <key>=<path>",
-			Value:   &app.secretFiles,
-		},
-		&cli.BoolFlag{
-			Name:        "artifact",
-			Aliases:     []string{"a"},
-			Usage:       "Output specified artifact; a wildcard (*) can be used to output all artifacts",
-			Destination: &app.artifactMode,
-		},
-		&cli.BoolFlag{
-			Name:        "image",
-			Usage:       "Output only docker image of the specified target",
-			Destination: &app.imageMode,
-		},
-		&cli.BoolFlag{
-			Name:        "pull",
-			EnvVars:     []string{"EARTHLY_PULL"},
-			Usage:       "Force pull any referenced Docker images",
-			Destination: &app.pull,
-		},
-		&cli.BoolFlag{
-			Name:        "push",
-			EnvVars:     []string{"EARTHLY_PUSH"},
-			Usage:       "Push docker images and execute RUN --push commands",
-			Destination: &app.push,
-		},
-		&cli.BoolFlag{
-			Name:        "ci",
-			EnvVars:     []string{"EARTHLY_CI"},
-			Usage:       "Execute in CI mode (implies --use-inline-cache --save-inline-cache --no-output --strict)",
-			Destination: &app.ci,
-		},
-		&cli.BoolFlag{
-			Name:        "output",
-			EnvVars:     []string{"EARTHLY_OUTPUT"},
-			Usage:       "Allow artifacts or images to be output, even when running under --ci mode",
-			Destination: &app.output,
-		},
-		&cli.BoolFlag{
-			Name:        "no-output",
-			EnvVars:     []string{"EARTHLY_NO_OUTPUT"},
-			Usage:       wrap("Do not output artifacts or images", "(using --push is still allowed)"),
-			Destination: &app.noOutput,
-		},
-		&cli.BoolFlag{
-			Name:        "no-cache",
-			EnvVars:     []string{"EARTHLY_NO_CACHE"},
-			Usage:       "Do not use cache while building",
-			Destination: &app.noCache,
-		},
-		&cli.StringFlag{
-			Name:        "config",
-			Value:       defaultConfigPath(),
-			EnvVars:     []string{"EARTHLY_CONFIG"},
-			Usage:       "Path to config file",
-			Destination: &app.configPath,
-		},
-		&cli.StringFlag{
-			Name:        "ssh-auth-sock",
-			Value:       os.Getenv("SSH_AUTH_SOCK"),
-			EnvVars:     []string{"EARTHLY_SSH_AUTH_SOCK"},
-			Usage:       wrap("The SSH auth socket to use for ssh-agent forwarding", ""),
-			Destination: &app.sshAuthSock,
-		},
-		&cli.StringFlag{
-			Name:        "auth-token",
-			EnvVars:     []string{"EARTHLY_TOKEN"},
-			Usage:       "Force Earthly account login to authenticate with supplied token",
-			Destination: &app.authToken,
-		},
-		&cli.StringFlag{
-			Name:        "git-username",
-			EnvVars:     []string{"GIT_USERNAME"},
-			Usage:       "The git username to use for git HTTPS authentication",
-			Destination: &app.gitUsernameOverride,
-		},
-		&cli.StringFlag{
-			Name:        "git-password",
-			EnvVars:     []string{"GIT_PASSWORD"},
-			Usage:       "The git password to use for git HTTPS authentication",
-			Destination: &app.gitPasswordOverride,
-		},
-		&cli.BoolFlag{
-			Name:        "allow-privileged",
-			Aliases:     []string{"P"},
-			EnvVars:     []string{"EARTHLY_ALLOW_PRIVILEGED"},
-			Usage:       "Allow build to use the --privileged flag in RUN commands",
-			Destination: &app.allowPrivileged,
-		},
-		&cli.BoolFlag{
-			Name:        "profiler",
-			EnvVars:     []string{"EARTHLY_PROFILER"},
-			Usage:       "Enable the profiler",
-			Destination: &app.enableProfiler,
-			Hidden:      true, // Dev purposes only.
-		},
-		&cli.StringFlag{
-			Name:        "buildkit-host",
-			Value:       "",
-			EnvVars:     []string{"EARTHLY_BUILDKIT_HOST"},
-			Usage:       wrap("The URL to use for connecting to a buildkit host. ", "If empty, earthly will attempt to start a buildkitd instance via docker run"),
-			Destination: &app.buildkitHost,
-		},
-		&cli.StringFlag{
-			Name:        "org",
-			EnvVars:     []string{"EARTHLY_ORG"},
-			Usage:       "The name of the organization the satellite belongs to. Required when using --satellite and user is a member of multiple organizations.",
-			Required:    false,
-			Destination: &app.satelliteOrg,
-		},
-		&cli.StringFlag{
-			Name:        "satellite",
-			Aliases:     []string{"sat"},
-			EnvVars:     []string{"EARTHLY_SATELLITE"},
-			Usage:       "The name of satellite to use for this build.",
-			Required:    false,
-			Destination: &app.satelliteName,
-		},
-		&cli.BoolFlag{
-			Name:        "no-satellite",
-			Aliases:     []string{"no-sat"},
-			EnvVars:     []string{"EARTHLY_NO_SATELLITE"},
-			Usage:       "Disables the use of a selected satellite for this build.",
-			Required:    false,
-			Destination: &app.noSatellite,
-		},
-		&cli.StringFlag{
-			Name:        "debugger-host",
-			EnvVars:     []string{"EARTHLY_DEBUGGER_HOST"},
-			Usage:       wrap("The URL to use for connecting to a debugger host. ", "If empty, earthly uses the default debugger port, combined with the desired buildkit host."),
-			Destination: &app.debuggerHost,
-			Hidden:      true,
-		},
-		&cli.StringFlag{
-			Name:        "tlscert",
-			Value:       "./certs/earthly_cert.pem",
-			EnvVars:     []string{"EARTHLY_TLS_CERT"},
-			Usage:       wrap("The path to the client TLS cert", "If relative, will be interpreted as relative to the ~/.earthly folder."),
-			Destination: &app.certPath,
-			Hidden:      true,
-		},
-		&cli.StringFlag{
-			Name:        "tlskey",
-			Value:       "./certs/earthly_key.pem",
-			EnvVars:     []string{"EARTHLY_TLS_KEY"},
-			Usage:       wrap("The path to the client TLS key.", "If relative, will be interpreted as relative to the ~/.earthly folder."),
-			Destination: &app.keyPath,
-			Hidden:      true,
-		},
-		&cli.StringFlag{
-			Name:        "buildkit-image",
-			Value:       DefaultBuildkitdImage,
-			EnvVars:     []string{"EARTHLY_BUILDKIT_IMAGE"},
-			Usage:       "The docker image to use for the buildkit daemon",
-			Destination: &app.buildkitdImage,
-		},
-		&cli.StringFlag{
-			Name:        "buildkit-container-name",
-			Value:       DefaultBuildkitdContainerName,
-			EnvVars:     []string{"EARTHLY_CONTAINER_NAME"},
-			Usage:       "The docker container name to use for the buildkit daemon",
-			Destination: &app.containerName,
-			Hidden:      true,
-		},
-		&cli.StringFlag{
-			Name:        "buildkit-volume-name",
-			Value:       DefaultBuildkitdVolumeName,
-			EnvVars:     []string{"EARTHLY_VOLUME_NAME"},
-			Usage:       "The docker volume name to use for the buildkit daemon cache",
-			Destination: &app.buildkitdSettings.VolumeName,
-			Hidden:      true,
-		},
-		&cli.StringFlag{
-			Name:        "remote-cache",
-			EnvVars:     []string{"EARTHLY_REMOTE_CACHE"},
-			Usage:       "A remote docker image tag use as explicit cache",
-			Destination: &app.remoteCache,
-		},
-		&cli.BoolFlag{
-			Name:        "max-remote-cache",
-			EnvVars:     []string{"EARTHLY_MAX_REMOTE_CACHE"},
-			Usage:       "Saves all intermediate images too in the remote cache",
-			Destination: &app.maxRemoteCache,
-		},
-		&cli.BoolFlag{
-			Name:        "save-inline-cache",
-			EnvVars:     []string{"EARTHLY_SAVE_INLINE_CACHE"},
-			Usage:       "Enable cache inlining when pushing images",
-			Destination: &app.saveInlineCache,
-		},
-		&cli.BoolFlag{
-			Name:        "use-inline-cache",
-			EnvVars:     []string{"EARTHLY_USE_INLINE_CACHE"},
-			Usage:       wrap("Attempt to use any inline cache that may have been previously pushed ", "uses image tags referenced by SAVE IMAGE --push or SAVE IMAGE --cache-from"),
-			Destination: &app.useInlineCache,
-		},
-		&cli.BoolFlag{
-			Name:        "interactive",
-			Aliases:     []string{"i"},
-			EnvVars:     []string{"EARTHLY_INTERACTIVE"},
-			Usage:       "Enable interactive debugging",
-			Destination: &app.interactiveDebugging,
-		},
-		&cli.BoolFlag{
-			Name:        "verbose",
-			Aliases:     []string{"V"},
-			EnvVars:     []string{"EARTHLY_VERBOSE"},
-			Usage:       "Enable verbose logging",
-			Destination: &app.verbose,
-		},
-		&cli.BoolFlag{
-			Name:        "debug",
-			Aliases:     []string{"D"},
-			EnvVars:     []string{"EARTHLY_DEBUG"},
-			Usage:       "Enable debug mode. This flag also turns on the debug mode of buildkitd, which may cause it to restart",
-			Destination: &app.debug,
-			Hidden:      true, // For development purposes only.
-		},
-		&cli.StringFlag{
-			Name:        "server",
-			Value:       "https://api.earthly.dev",
-			EnvVars:     []string{"EARTHLY_SERVER_ADDRESS"},
-			Usage:       "API server override for dev purposes",
-			Destination: &app.apiServer,
-			Hidden:      true, // Internal.
-		},
-		&cli.StringFlag{
-			Name:        "satellite-address",
-			Value:       containerutil.SatelliteAddress,
-			EnvVars:     []string{"EARTHLY_SATELLITE_ADDRESS"},
-			Usage:       "Satellite address override for dev purposes",
-			Destination: &app.satelliteAddress,
-			Hidden:      true, // Internal.
-		},
-		&cli.BoolFlag{
-			Name:        "no-fake-dep",
-			EnvVars:     []string{"EARTHLY_NO_FAKE_DEP"},
-			Usage:       "Internal feature flag for fake-dep",
-			Destination: &app.noFakeDep,
-			Hidden:      true, // Internal.
-		},
-		&cli.BoolFlag{
-			Name:        "strict",
-			EnvVars:     []string{"EARTHLY_STRICT"},
-			Usage:       "Disallow usage of features that may create unrepeatable builds",
-			Destination: &app.strict,
-		},
-		// TODO: completely remove conversion-parallelism in some future release
-		&cli.IntFlag{
-			Name:        "conversion-parallelism",
-			EnvVars:     []string{"EARTHLY_CONVERSION_PARALLELISM"},
-			Usage:       "This flag is obsolete, use 'earthly config global.conversion_parallelism <parallelism>' instead'",
-			Destination: &app.conversionParllelism,
-			Hidden:      true, // obsolete in favor of config
-		},
-		&cli.BoolFlag{
-			EnvVars:     []string{"EARTHLY_DISABLE_ANALYTICS", "DO_NOT_TRACK"},
-			Usage:       "Disable collection of analytics",
-			Destination: &app.disableAnalytics,
-		},
-		&cli.StringFlag{
-			Name:        "version-flag-overrides",
-			EnvVars:     []string{"EARTHLY_VERSION_FLAG_OVERRIDES"},
-			Usage:       "Apply additional flags after each VERSION command across all Earthfiles, multiple flags can be seperated by commas",
-			Destination: &app.featureFlagOverrides,
-			Hidden:      true, // used for feature-flipping from ./earthly dev script
-		},
-		&cli.StringFlag{
-			Name:        envFileFlag,
-			EnvVars:     []string{"EARTHLY_ENV_FILE"},
-			Usage:       "Use values from this file as earthly environment variables, buildargs, or secrets",
-			Value:       defaultEnvFile,
-			Destination: &app.envFile,
-		},
-	}
 
-	app.cliApp.Commands = []*cli.Command{
-		{
-			Name:        "bootstrap",
-			Usage:       "Bootstraps earthly installation including shell autocompletion and buildkit image download",
-			Description: "Bootstraps earthly installation including shell autocompletion and buildkit image download",
-			Action:      app.actionBootstrap,
-			Flags: []cli.Flag{
-				&cli.StringFlag{
-					Name:        "source",
-					Usage:       "Output source file (for use in homebrew install)",
-					Hidden:      true, // only meant for use with homebrew formula
-					Destination: &app.homebrewSource,
-				},
-				&cli.BoolFlag{
-					Name:        "no-buildkit",
-					Usage:       "Do not bootstrap buildkit",
-					Destination: &app.bootstrapNoBuildkit,
-				},
-				&cli.BoolFlag{
-					Name:        "with-autocomplete",
-					Usage:       "Add earthly autocompletions",
-					Destination: &app.bootstrapWithAutocomplete,
-				},
-			},
-		},
-		{
-			Name:        "docker",
-			Usage:       "Build a Dockerfile without converting to an Earthfile *experimental*",
-			Description: "Builds a dockerfile",
-			Hidden:      true, // Experimental.
-			Action:      app.actionDocker,
-			Flags: []cli.Flag{
-				&cli.StringFlag{
-					Name:        "dockerfile",
-					Usage:       "Path to dockerfile input, or - for stdin",
-					Value:       "Dockerfile",
-					Destination: &app.dockerfilePath,
-				},
-				&cli.StringFlag{
-					Name:        "tag",
-					Usage:       "Name and tag for the built image; formatted as 'name:tag'",
-					Destination: &app.earthfileFinalImage,
-				},
-			},
-		},
-		{
-			Name:        "docker2earthly",
-			Usage:       "Convert a Dockerfile into Earthfile",
-			Description: "Converts an existing dockerfile into an Earthfile",
-			Hidden:      true, // Experimental.
-			Action:      app.actionDocker2Earthly,
-			Flags: []cli.Flag{
-				&cli.StringFlag{
-					Name:        "dockerfile",
-					Usage:       "Path to dockerfile input, or - for stdin",
-					Value:       "Dockerfile",
-					Destination: &app.dockerfilePath,
-				},
-				&cli.StringFlag{
-					Name:        "earthfile",
-					Usage:       "Path to Earthfile output, or - for stdout",
-					Value:       "Earthfile",
-					Destination: &app.earthfilePath,
-				},
-				&cli.StringFlag{
-					Name:        "tag",
-					Usage:       "Name and tag for the built image; formatted as 'name:tag'",
-					Destination: &app.earthfileFinalImage,
-				},
-			},
-		},
-		{
-			Name:    "org",
-			Aliases: []string{"orgs"},
-			Usage:   "Earthly organization administration *experimental*",
-			Subcommands: []*cli.Command{
-				{
-					Name:      "create",
-					Usage:     "Create a new organization",
-					UsageText: "earthly [options] org create <org-name>",
-					Action:    app.actionOrgCreate,
-				},
-				{
-					Name:      "ls",
-					Aliases:   []string{"list"},
-					Usage:     "List organizations you belong to",
-					UsageText: "earthly [options] org ls",
-					Action:    app.actionOrgList,
-				},
-				{
-					Name:      "list-permissions",
-					Usage:     "List permissions and membership of an organization",
-					UsageText: "earthly [options] org list-permissions <org-name>",
-					Action:    app.actionOrgListPermissions,
-				},
-				{
-					Name:      "invite",
-					Usage:     "Invite accounts to your organization",
-					UsageText: "earthly [options] org invite [options] <path> <email> [<email> ...]",
-					Action:    app.actionOrgInvite,
-					Flags: []cli.Flag{
-						&cli.BoolFlag{
-							Name:        "write",
-							Usage:       "Grant write permissions in addition to read",
-							Destination: &app.writePermission,
-						},
-					},
-				},
-				{
-					Name:      "revoke",
-					Usage:     "Remove accounts from your organization",
-					UsageText: "earthly [options] org revoke <path> <email> [<email> ...]",
-					Action:    app.actionOrgRevoke,
-				},
-			},
-		},
-		{
-			Name:      "ls",
-			Usage:     "List targets from an Earthfile *experimental*",
-			UsageText: "earthly [options] ls [<project-ref>]",
-			Action:    app.actionListTargets,
-			Flags: []cli.Flag{
-				&cli.BoolFlag{
-					Name:        "args",
-					Aliases:     []string{"a"},
-					Usage:       "Show Arguments",
-					Destination: &app.lsShowArgs,
-				},
-				&cli.BoolFlag{
-					Name:        "long",
-					Aliases:     []string{"l"},
-					Usage:       "Show full target-ref",
-					Destination: &app.lsShowLong,
-				},
-			},
-		},
-		{
-			Name:        "secret",
-			Aliases:     []string{"secrets"},
-			Usage:       "Earthly secrets",
-			Description: "Manage cloud secrets *experimental*",
-			Subcommands: []*cli.Command{
-				{
-					Name:  "set",
-					Usage: "Stores a secret in the secrets store",
-					UsageText: "earthly [options] secret set <path> <value>\n" +
-						"   earthly [options] secret set --file <local-path> <path>\n" +
-						"   earthly [options] secret set --file <local-path> <path>",
-					Action: app.actionSecretsSet,
-					Flags: []cli.Flag{
-						&cli.StringFlag{
-							Name:        "file",
-							Aliases:     []string{"f"},
-							Usage:       "Stores secret stored in file",
-							Destination: &app.secretFile,
-						},
-						&cli.BoolFlag{
-							Name:        "stdin",
-							Aliases:     []string{"i"},
-							Usage:       "Stores secret read from stdin",
-							Destination: &app.secretStdin,
-						},
-					},
-				},
-				{
-					Name:      "get",
-					Action:    app.actionSecretsGet,
-					Usage:     "Retrieve a secret from the secrets store",
-					UsageText: "earthly [options] secret get [options] <path>",
-					Flags: []cli.Flag{
-						&cli.BoolFlag{
-							Aliases:     []string{"n"},
-							Usage:       "Disable newline at the end of the secret",
-							Destination: &app.disableNewLine,
-						},
-					},
-				},
-				{
-					Name:      "ls",
-					Usage:     "List secrets in the secrets store",
-					UsageText: "earthly [options] secret ls [<path>]",
-					Action:    app.actionSecretsList,
-				},
-				{
-					Name:      "rm",
-					Usage:     "Removes a secret from the secrets store",
-					UsageText: "earthly [options] secret rm <path>",
-					Action:    app.actionSecretsRemove,
-				},
-			},
-		},
-		{
-			Name:  "account",
-			Usage: "Create or manage an Earthly account *experimental*",
-			Subcommands: []*cli.Command{
-				{
-					Name:        "register",
-					Usage:       "Register for an Earthly account",
-					Description: "Register for an Earthly account",
-					UsageText: "You may register using GitHub OAuth, by visiting https://ci.earthly.dev\n" +
-						"   Once authenticated, a login token will be displayed which can be used to login:\n" +
-						"\n" +
-						"       earthly [options] account login --token <token>\n" +
-						"\n" +
-						"   Alternatively, you can register using an email:\n" +
-						"       first, request a token with:\n" +
-						"\n" +
-						"           earthly [options] account register --email <email>\n" +
-						"\n" +
-						"       then check your email to retrieve the token, then continue by running:\n" +
-						"\n" +
-						"           earthly [options] account register --token <token>\n",
-					Action: app.actionRegister,
-					Flags: []cli.Flag{
-						&cli.StringFlag{
-							Name:        "email",
-							Usage:       "Email address",
-							Destination: &app.email,
-						},
-						&cli.StringFlag{
-							Name:        "token",
-							Usage:       "Email verification token",
-							Destination: &app.token,
-						},
-						&cli.StringFlag{
-							Name:        "password",
-							EnvVars:     []string{"EARTHLY_PASSWORD"},
-							Usage:       "Specify password on the command line instead of interactively being asked",
-							Destination: &app.password,
-						},
-						&cli.StringFlag{
-							Name:        "public-key",
-							EnvVars:     []string{"EARTHLY_PUBLIC_KEY"},
-							Usage:       "Path to public key to register",
-							Destination: &app.registrationPublicKey,
-						},
-						&cli.BoolFlag{
-							Name:        "accept-terms-of-service-privacy",
-							EnvVars:     []string{"EARTHLY_ACCEPT_TERMS_OF_SERVICE_PRIVACY"},
-							Usage:       "Accept the Terms & Conditions, and Privacy Policy",
-							Destination: &app.termsConditionsPrivacy,
-						},
-					},
-				},
-				{
-					Name:        "login",
-					Usage:       "Login to an Earthly account",
-					Description: "Login to an Earthly account",
-					UsageText: "earthly [options] account login\n" +
-						"   earthly [options] account login --email <email>\n" +
-						"   earthly [options] account login --email <email> --password <password>\n" +
-						"   earthly [options] account login --token <token>\n",
-					Action: app.actionAccountLogin,
-					Flags: []cli.Flag{
-						&cli.StringFlag{
-							Name:        "email",
-							Usage:       "Email address",
-							Destination: &app.email,
-						},
-						&cli.StringFlag{
-							Name:        "token",
-							Usage:       "Authentication token",
-							Destination: &app.token,
-						},
-						&cli.StringFlag{
-							Name:        "password",
-							EnvVars:     []string{"EARTHLY_PASSWORD"},
-							Usage:       "Specify password on the command line instead of interactively being asked",
-							Destination: &app.password,
-						},
-					},
-				},
-				{
-					Name:        "logout",
-					Usage:       "Logout of an Earthly account",
-					Description: "Logout of an Earthly account; this has no effect for ssh-based authentication",
-					Action:      app.actionAccountLogout,
-				},
-				{
-					Name:      "list-keys",
-					Usage:     "List associated public keys used for authentication",
-					UsageText: "earthly [options] account list-keys",
-					Action:    app.actionAccountListKeys,
-				},
-				{
-					Name:      "add-key",
-					Usage:     "Associate a new public key with your account",
-					UsageText: "earthly [options] add-key [<key>]",
-					Action:    app.actionAccountAddKey,
-				},
-				{
-					Name:      "remove-key",
-					Usage:     "Removes an existing public key from your account",
-					UsageText: "earthly [options] remove-key <key>",
-					Action:    app.actionAccountRemoveKey,
-				},
-				{
-					Name:      "list-tokens",
-					Usage:     "List associated tokens used for authentication",
-					UsageText: "earthly [options] account list-tokens",
-					Action:    app.actionAccountListTokens,
-				},
-				{
-					Name:      "create-token",
-					Usage:     "Create a new authentication token for your account",
-					UsageText: "earthly [options] account create-token [options] <token name>",
-					Action:    app.actionAccountCreateToken,
-					Flags: []cli.Flag{
-						&cli.BoolFlag{
-							Name:        "write",
-							Usage:       "Grant write permissions in addition to read",
-							Destination: &app.writePermission,
-						},
-						&cli.StringFlag{
-							Name:        "expiry",
-							Usage:       "Set token expiry date in the form YYYY-MM-DD or never (default 1year)",
-							Destination: &app.expiry,
-						},
-					},
-				},
-				{
-					Name:      "remove-token",
-					Usage:     "Remove an authentication token from your account",
-					UsageText: "earthly [options] account remove-token <token>",
-					Action:    app.actionAccountRemoveToken,
-				},
-			},
-		},
-		{
-			Name:        "debug",
-			Usage:       "Print debug information about an Earthfile",
-			Description: "Print debug information about an Earthfile",
-			ArgsUsage:   "[<path>]",
-			Hidden:      true, // Dev purposes only.
-			Subcommands: []*cli.Command{
-				{
-					Name:      "ast",
-					Usage:     "Output the AST",
-					UsageText: "earthly [options] debug ast",
-					Action:    app.actionDebugAst,
-					Flags: []cli.Flag{
-						&cli.BoolFlag{
-							Name:        "source-map",
-							Usage:       "Enable outputting inline sourcemap",
-							Destination: &app.enableSourceMap,
-						},
-					},
-				},
-			},
-		},
-		{
-			Name:        "prune",
-			Usage:       "Prune Earthly build cache",
-			Description: "Prune Earthly build cache",
-			Action:      app.actionPrune,
-			Flags: []cli.Flag{
-				&cli.BoolFlag{
-					Name:        "all",
-					Aliases:     []string{"a"},
-					EnvVars:     []string{"EARTHLY_PRUNE_ALL"},
-					Usage:       "Prune all cache",
-					Destination: &app.pruneAll,
-				},
-				&cli.BoolFlag{
-					Name:        "reset",
-					EnvVars:     []string{"EARTHLY_PRUNE_RESET"},
-					Usage:       "Reset cache entirely by wiping cache dir",
-					Destination: &app.pruneReset,
-				},
-			},
-		},
-		{
-			Name:   "config",
-			Usage:  "Edits your Earthly configuration file",
-			Action: app.actionConfig,
-			UsageText: `This command takes a path, and a value and sets it in your configuration file.
+	app.cliApp.Flags = app.rootFlags()                                     // These will show up in help.
+	app.cliApp.Flags = append(app.cliApp.Flags, app.hiddenBuildFlags()...) // These will not.
 
-	 As the configuration file is YAML, the key must be a valid key within the file. You can specify sub-keys by using "." to separate levels.
-	 If the sub-key you wish to use has a "." in it, you can quote that subsection, like this: git."github.com".
-
-	 Values must be valid YAML, and also be deserializable into the key you wish to assign them to.
-	 This means you can set higher level objects using a compact style, or single values with simple values.
-
-	 Only one key/value can be set per invocation.
-
-	 To get help with a specific key, do "config [key] --help". Or, visit https://docs.earthly.dev/earthly-config for more details.`,
-			Description: `Set your cache size:
-
-	config global.cache_size_mb 1234
-
-Set additional buildkit args, using a YAML array:
-
-	config global.buildkit_additional_args '["userns", "--host"]'
-
-Set a key containing a period:
-
-	config 'git."example.com".password' hunter2
-
-Set up a whole custom git repository for a server called example.com, using a single-line YAML literal:
-	* which stores git repos under /var/git/repos/name-of-repo.git
-	* allows access over ssh
-	* using port 2222
-	* sets the username to git
-	* is recognized to earthly as example.com/name-of-repo
-
-	config git "{example: {pattern: 'example.com/([^/]+)', substitute: 'ssh://git@example.com:2222/var/git/repos/\$1.git', auth: ssh}}"
-			`,
-			Flags: []cli.Flag{
-				&cli.BoolFlag{
-					Name:        "dry-run",
-					Usage:       "Print the changed config file to the console instead of writing it out",
-					Destination: &app.configDryRun,
-				},
-			},
-		},
-		{
-			Name:    "satellite",
-			Aliases: []string{"satellites", "sat"},
-			Usage: "Launch and use a Satellite runner as remote backend for Earthly builds.\n" +
-				"	Satellites can be used to optimize and share cache between multiple builds and users,\n" +
-				"	as well as run builds in native architectures independent of where the Earthly client is invoked.\n" +
-				"	Note: this feature is currently experimental.\n" +
-				"	If you'd like to try it out, please contact us via Slack to be added to the beta testers group.",
-			UsageText:   "earthly satellite (launch|ls|inspect|select|unselect|rm)",
-			Description: "Create and manage Earthly Satellites",
-			Flags: []cli.Flag{
-				&cli.StringFlag{
-					Name:        "org",
-					EnvVars:     []string{"EARTHLY_ORG"},
-					Usage:       "The name of the organization the satellite belongs to. Required when user is a member of multiple.",
-					Required:    false,
-					Destination: &app.satelliteOrg,
-				},
-			},
-			Subcommands: []*cli.Command{
-				{
-					Name:        "launch",
-					Usage:       "Launch a new Earthly Satellite",
-					Description: "Launch a new Earthly Satellite",
-					UsageText: "earthly satellite launch <satellite-name>\n" +
-						"	earthly satellite --org <organization-name> launch <satellite-name>",
-					Action: app.actionSatelliteLaunch,
-				},
-				{
-					Name:        "rm",
-					Usage:       "Destroy an Earthly Satellite",
-					Description: "Destroy an Earthly Satellite",
-					UsageText: "earthly satellite rm <satellite-name>\n" +
-						"	earthly satellite --org <organization-name> rm <satellite-name>",
-					Action: app.actionSatelliteDestroy,
-				},
-				{
-					Name:        "ls",
-					Description: "List your Earthly Satellites",
-					Usage:       "List your Earthly Satellites",
-					UsageText: "earthly satellite ls\n" +
-						"	earthly satellite --org <organization-name> ls",
-					Action: app.actionSatelliteList,
-				},
-				{
-					Name:        "inspect",
-					Description: "Show additional details about a Satellite instance",
-					Usage:       "Show additional details about a Satellite instance",
-					UsageText: "earthly satellite inspect <satellite-name>\n" +
-						"	earthly satellite --org <organization-name> inspect <satellite-name>",
-					Action: app.actionSatelliteDescribe,
-				},
-				{
-					Name:        "select",
-					Aliases:     []string{"s"},
-					Usage:       "Choose which satellite to use to build your app",
-					Description: "Choose which satellite to use to build your app",
-					UsageText: "earthly satellite select <satellite-name>\n" +
-						"	earthly satellite --org <organization-name> select <satellite-name>",
-					Action: app.actionSatelliteSelect,
-				},
-				{
-					Name:        "unselect",
-					Aliases:     []string{"uns"},
-					Usage:       "Remove any currently selected Satellite instance from your Earthly configuration",
-					Description: "Remove any currently selected Satellite instance from your Earthly configuration",
-					UsageText:   "earthly satellite unselect",
-					Action:      app.actionSatelliteUnselect,
-				},
-			},
-		},
-	}
+	app.cliApp.Commands = app.rootCmds()
 
 	app.cliApp.Before = app.before
 	return app
-}
-
-func wrap(s ...string) string {
-	return strings.Join(s, "\n\t")
 }
 
 func (app *earthlyApp) before(context *cli.Context) error {
@@ -1317,105 +489,6 @@ func (app *earthlyApp) before(context *cli.Context) error {
 	return nil
 }
 
-func (app *earthlyApp) configureSatellite(cliCtx *cli.Context, cloudClient cloud.Client) error {
-	if cliCtx.IsSet("buildkit-host") && cliCtx.IsSet("satellite") {
-		return errors.New("cannot specify both buildkit-host and satellite")
-	}
-	if cliCtx.IsSet("satellite") && app.noSatellite {
-		return errors.New("cannot specify both no-satellite and satellite")
-	}
-	if !app.isUsingSatellite(cliCtx) || cloudClient == nil {
-		// If the app is not using a cloud client, or the command doesn't interact with the cloud (prune, bootstrap)
-		// then pretend its all good and use your regular configuration.
-		return nil
-	}
-
-	// When using a satellite, interactive and local do not work; as they are not SSL nor routable yet.
-	app.console.Warnf("Note: the interactive debugger, interactive RUN commands, and fast output via embedded registry do not yet work on Earthly Satellites.")
-
-	// Set up extra settings needed for buildkit RPC metadata
-	if app.satelliteName == "" {
-		app.satelliteName = app.cfg.Satellite.Name
-	}
-	if app.satelliteOrg == "" {
-		app.satelliteOrg = app.cfg.Satellite.Org
-	}
-	orgID, err := app.getSatelliteOrgID(cliCtx.Context, cloudClient)
-	if err != nil {
-		return err
-	}
-	app.buildkitdSettings.SatelliteName = app.satelliteName
-	app.buildkitdSettings.SatelliteOrgID = orgID
-	if app.satelliteAddress != "" {
-		app.buildkitdSettings.BuildkitAddress = app.satelliteAddress
-	} else {
-		app.buildkitdSettings.BuildkitAddress = containerutil.SatelliteAddress
-	}
-	app.analyticsMetadata.isSatellite = true
-	app.analyticsMetadata.satelliteVersion = "" // TODO
-
-	app.console.Warnf("") // newline
-	app.console.Warnf("The following feature flags are recommended for use with Satellites and will be auto-enabled:")
-	app.console.Warnf("--new-platform, --use-registry-for-with-docker")
-	app.console.Warnf("") // newline
-
-	if app.featureFlagOverrides != "" {
-		app.featureFlagOverrides += ","
-	}
-	app.featureFlagOverrides += "new-platform,use-registry-for-with-docker"
-
-	token, err := cloudClient.GetAuthToken(cliCtx.Context)
-	if err != nil {
-		return errors.Wrap(err, "failed to get auth token")
-	}
-	app.buildkitdSettings.SatelliteToken = token
-
-	// TODO (dchw) what other settings might we want to override here?
-	return nil
-}
-
-func (app *earthlyApp) isUsingSatellite(cliCtx *cli.Context) bool {
-	if app.noSatellite {
-		return false
-	}
-	if cliCtx.IsSet("buildkit-host") {
-		// buildkit-host takes precedence
-		return false
-	}
-	return app.cfg.Satellite.Name != "" || app.satelliteName != ""
-}
-
-func (app *earthlyApp) GetBuildkitClient(cliCtx *cli.Context, cloudClient cloud.Client) (*client.Client, error) {
-	err := app.configureSatellite(cliCtx, cloudClient)
-	if err != nil {
-		return nil, errors.Wrapf(err, "could not construct new buildkit client")
-	}
-
-	return buildkitd.NewClient(cliCtx.Context, app.console, app.buildkitdImage, app.containerName, app.containerFrontend, app.buildkitdSettings)
-}
-
-func (app *earthlyApp) handleTLSCertificateSettings(context *cli.Context) {
-	if !app.cfg.Global.TLSEnabled {
-		return
-	}
-
-	app.buildkitdSettings.TLSCA = app.cfg.Global.TLSCA
-
-	if !context.IsSet("tlscert") && app.cfg.Global.ClientTLSCert != "" {
-		app.certPath = app.cfg.Global.ClientTLSCert
-	}
-
-	if !context.IsSet("tlskey") && app.cfg.Global.ClientTLSKey != "" {
-		app.keyPath = app.cfg.Global.ClientTLSKey
-	}
-
-	app.buildkitdSettings.ClientTLSCert = app.certPath
-	app.buildkitdSettings.ClientTLSKey = app.keyPath
-
-	app.buildkitdSettings.ServerTLSCert = app.cfg.Global.ServerTLSCert
-	app.buildkitdSettings.ServerTLSKey = app.cfg.Global.ServerTLSKey
-}
-
 func (app *earthlyApp) warnIfEarth() {
 	if len(os.Args) == 0 {
 		return
@@ -1513,215 +586,6 @@ func unhideFlagsCommands(ctx context.Context, cmds []*cli.Command) {
 		}
 		unhideFlagsCommands(ctx, cmd.Subcommands)
 	}
-}
-
-// to enable autocomplete, enter
-// complete -o nospace -C "/path/to/earthly" earthly
-func (app *earthlyApp) autoComplete(ctx context.Context) {
-	_, found := os.LookupEnv("COMP_LINE")
-	if !found {
-		return
-	}
-
-	app.console = app.console.WithLogLevel(conslogging.Silent)
-
-	err := app.autoCompleteImp(ctx)
-	if err != nil {
-		errToLog := err
-		logDir, err := cliutil.GetOrCreateEarthlyDir()
-		if err != nil {
-			os.Exit(1)
-		}
-		logFile := filepath.Join(logDir, "autocomplete.log")
-		err = os.MkdirAll(logDir, 0755)
-		if err != nil {
-			os.Exit(1)
-		}
-		f, err := os.OpenFile(logFile, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0755)
-		if err != nil {
-			os.Exit(1)
-		}
-		fmt.Fprintf(f, "error during autocomplete: %s\n", errToLog)
-		os.Exit(1)
-	}
-	os.Exit(0)
-}
-
-func (app *earthlyApp) autoCompleteImp(ctx context.Context) (err error) {
-	defer func() {
-		if r := recover(); r != nil {
-			err = errors.Errorf("recovered panic in autocomplete %s: %s", r, debug.Stack())
-		}
-	}()
-
-	compLine := os.Getenv("COMP_LINE")   // full command line
-	compPoint := os.Getenv("COMP_POINT") // where the cursor is
-
-	compPointInt, err := strconv.ParseUint(compPoint, 10, 64)
-	if err != nil {
-		return err
-	}
-
-	gitLookup := buildcontext.NewGitLookup(app.console, app.sshAuthSock)
-	resolver := buildcontext.NewResolver("", nil, gitLookup, app.console, "")
-	var gwClient gwclient.Client // TODO this is a nil pointer which causes a panic if we try to expand a remotely referenced earthfile
-	// it's expensive to create this gwclient, so we need to implement a lazy eval which returns it when required.
-
-	potentials, err := autocomplete.GetPotentials(ctx, resolver, gwClient, compLine, int(compPointInt), app.cliApp)
-	if err != nil {
-		return err
-	}
-	for _, p := range potentials {
-		fmt.Printf("%s\n", p)
-	}
-
-	return err
-}
-func (app *earthlyApp) insertBashCompleteEntry() error {
-	var path string
-	if runtime.GOOS == "darwin" {
-		path = "/usr/local/etc/bash_completion.d/earthly"
-	} else {
-		path = "/usr/share/bash-completion/completions/earthly"
-	}
-	dirPath := filepath.Dir(path)
-
-	dirPathExists, err := fileutil.DirExists(dirPath)
-	if err != nil {
-		return errors.Wrapf(err, "failed checking if %s exists", dirPath)
-	}
-	if !dirPathExists {
-		fmt.Fprintf(os.Stderr, "Warning: unable to enable bash-completion: %s does not exist\n", dirPath)
-		return nil // bash-completion isn't available, silently fail.
-	}
-
-	pathExists, err := fileutil.FileExists(path)
-	if err != nil {
-		return errors.Wrapf(err, "failed checking if %s exists", path)
-	}
-	if pathExists {
-		return nil // file already exists, don't update it.
-	}
-
-	// create the completion file
-	f, err := os.Create(path)
-	if err != nil {
-		return err
-	}
-	defer f.Close()
-
-	bashEntry, err := bashCompleteEntry()
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Warning: unable to enable bash-completion: %s\n", err)
-		return nil // bash-completion isn't available, silently fail.
-	}
-
-	_, err = f.Write([]byte(bashEntry))
-	if err != nil {
-		return errors.Wrapf(err, "failed writing to %s", path)
-	}
-	return nil
-}
-
-func (app *earthlyApp) deleteZcompdump() error {
-	var homeDir string
-	sudoUser, found := os.LookupEnv("SUDO_USER")
-	if !found {
-		var err error
-		homeDir, err = os.UserHomeDir()
-		if err != nil {
-			return errors.Wrapf(err, "failed to lookup current user home dir")
-		}
-	} else {
-		currentUser, err := user.Lookup(sudoUser)
-		if err != nil {
-			return errors.Wrapf(err, "failed to lookup user %s", sudoUser)
-		}
-		homeDir = currentUser.HomeDir
-	}
-	files, err := os.ReadDir(homeDir)
-	if err != nil {
-		return errors.Wrapf(err, "failed to read dir %s", homeDir)
-	}
-	for _, f := range files {
-		if strings.HasPrefix(f.Name(), ".zcompdump") {
-			path := filepath.Join(homeDir, f.Name())
-			err := os.Remove(path)
-			if err != nil {
-				return errors.Wrapf(err, "failed to remove %s", path)
-			}
-		}
-	}
-	return nil
-}
-
-func bashCompleteEntry() (string, error) {
-	template := "complete -o nospace -C '__earthly__' earthly\n"
-	return renderEntryTemplate(template)
-}
-
-func zshCompleteEntry() (string, error) {
-	template := `#compdef _earthly earthly
-
-function _earthly {
-    autoload -Uz bashcompinit
-    bashcompinit
-    complete -o nospace -C '__earthly__' earthly
-}
-`
-	return renderEntryTemplate(template)
-}
-
-func renderEntryTemplate(template string) (string, error) {
-	earthlyPath, err := os.Executable()
-	if err != nil {
-		return "", errors.Wrapf(err, "failed to determine earthly path: %s", err)
-	}
-	return strings.ReplaceAll(template, "__earthly__", earthlyPath), nil
-}
-
-// If debugging this, it might be required to run `rm ~/.zcompdump*` to remove the cache
-func (app *earthlyApp) insertZSHCompleteEntry() error {
-	// should be the same on linux and macOS
-	path := "/usr/local/share/zsh/site-functions/_earthly"
-	dirPath := filepath.Dir(path)
-
-	dirPathExists, err := fileutil.DirExists(dirPath)
-	if err != nil {
-		return errors.Wrapf(err, "failed to check if %s exists", dirPath)
-	}
-	if !dirPathExists {
-		fmt.Fprintf(os.Stderr, "Warning: unable to enable zsh-completion: %s does not exist\n", dirPath)
-		return nil // zsh-completion isn't available, silently fail.
-	}
-
-	pathExists, err := fileutil.FileExists(path)
-	if err != nil {
-		return errors.Wrapf(err, "failed to check if %s exists", path)
-	}
-	if pathExists {
-		return nil // file already exists, don't update it.
-	}
-
-	// create the completion file
-	f, err := os.Create(path)
-	if err != nil {
-		return err
-	}
-	defer f.Close()
-
-	compEntry, err := zshCompleteEntry()
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Warning: unable to enable zsh-completion: %s\n", err)
-		return nil // zsh-completion isn't available, silently fail.
-	}
-
-	_, err = f.Write([]byte(compEntry))
-	if err != nil {
-		return errors.Wrapf(err, "failed writing to %s", path)
-	}
-
-	return app.deleteZcompdump()
 }
 
 func (app *earthlyApp) run(ctx context.Context, args []string) int {
@@ -1840,1882 +704,4 @@ func (app *earthlyApp) printCrashLogs(ctx context.Context) {
 		app.console.PrintBar(color.New(color.FgHiRed), "Buildkit Logs", "")
 		fmt.Fprintln(os.Stderr, logs)
 	}
-}
-
-func isEarthlyBinary(path string) bool {
-	// apply heuristics to see if binary is a version of earthly
-	data, err := os.ReadFile(path)
-	if err != nil {
-		return false
-	}
-	if !bytes.Contains(data, []byte("docs.earthly.dev")) {
-		return false
-	}
-	if !bytes.Contains(data, []byte("api.earthly.dev")) {
-		return false
-	}
-	if !bytes.Contains(data, []byte("Earthfile")) {
-		return false
-	}
-	return true
-}
-
-func symlinkEarthlyToEarth() error {
-	binPath, err := os.Executable()
-	if err != nil {
-		return errors.Wrap(err, "failed to get current executable path")
-	}
-
-	baseName := path.Base(binPath)
-	if baseName != "earthly" {
-		return nil
-	}
-
-	earthPath := path.Join(path.Dir(binPath), "earth")
-
-	earthPathExists, err := fileutil.FileExists(earthPath)
-	if err != nil {
-		return errors.Wrapf(err, "failed to check if %s exists", earthPath)
-	}
-	if !earthPathExists && termutil.IsTTY() {
-		return nil // legacy earth binary doesn't exist, don't create it (unless we're under a non-tty system e.g. CI)
-	}
-
-	if !isEarthlyBinary(earthPath) {
-		return nil // file exists but is not an earthly binary, leave it alone.
-	}
-
-	// otherwise legacy earth command has been detected, remove it and symlink
-	// to the new earthly command.
-	err = os.Remove(earthPath)
-	if err != nil {
-		return errors.Wrapf(err, "failed to remove old install at %s", earthPath)
-	}
-	err = os.Symlink(binPath, earthPath)
-	if err != nil {
-		return errors.Wrapf(err, "failed to symlink %s to %s", binPath, earthPath)
-	}
-	return nil
-}
-
-func (app *earthlyApp) actionBootstrap(cliCtx *cli.Context) error {
-	app.commandName = "bootstrap"
-
-	switch app.homebrewSource {
-	case "bash":
-		compEntry, err := bashCompleteEntry()
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "Warning: failed to enable bash-completion: %s\n", err)
-			return nil // zsh-completion isn't available, silently fail.
-		}
-		fmt.Print(compEntry)
-		return nil
-	case "zsh":
-		compEntry, err := zshCompleteEntry()
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "Warning: failed to bootstrap zsh-completion: %s\n", err)
-			return nil // zsh-completion isn't available, silently fail.
-		}
-		fmt.Print(compEntry)
-		return nil
-	case "":
-		break
-	default:
-		return errors.Errorf("unhandled source %q", app.homebrewSource)
-	}
-
-	return app.bootstrap(cliCtx)
-}
-
-func (app *earthlyApp) bootstrap(cliCtx *cli.Context) error {
-	var err error
-	console := app.console.WithPrefix("bootstrap")
-	defer func() {
-		// cliutil.IsBootstrapped() determines if bootstrapping was done based
-		// on the existance of ~/.earthly; therefore we must ensure it's created.
-		cliutil.GetOrCreateEarthlyDir()
-		cliutil.EnsurePermissions()
-	}()
-
-	if app.bootstrapWithAutocomplete {
-		// Because this requires sudo, it should warn and not fail the rest of it.
-		err = app.insertBashCompleteEntry()
-		if err != nil {
-			console.Warnf("Warning: %s\n", err.Error())
-			err = nil
-		}
-		err = app.insertZSHCompleteEntry()
-		if err != nil {
-			console.Warnf("Warning: %s\n", err.Error())
-			err = nil
-		}
-
-		console.Printf("You may have to restart your shell for autocomplete to get initialized (e.g. run \"exec $SHELL\")\n")
-	}
-
-	err = symlinkEarthlyToEarth()
-	if err != nil {
-		console.Warnf("Warning: %s\n", err.Error())
-		err = nil
-	}
-
-	if !app.bootstrapNoBuildkit && !app.isUsingSatellite(cliCtx) {
-		bkURL, err := url.Parse(app.buildkitHost)
-		if err != nil {
-			return errors.Wrapf(err, "invalid buildkit_host: %s", app.cfg.Global.BuildkitHost)
-		}
-		if bkURL.Scheme == "tcp" && app.cfg.Global.TLSEnabled {
-			root, err := cliutil.GetOrCreateEarthlyDir()
-			if err != nil {
-				return err
-			}
-
-			certsDir := filepath.Join(root, "certs")
-			err = buildkitd.GenerateCertificates(certsDir)
-			if err != nil {
-				return errors.Wrap(err, "setup TLS")
-			}
-		}
-
-		// Bootstrap buildkit - pulls image and starts daemon.
-		bkClient, err := app.GetBuildkitClient(cliCtx, nil)
-		if err != nil {
-			return errors.Wrap(err, "bootstrap new buildkitd client")
-		}
-		defer bkClient.Close()
-	}
-
-	console.Printf("Bootstrapping successful.\n")
-	return nil
-}
-
-func promptInput(ctx context.Context, question string) (string, error) {
-	fmt.Printf("%s", question)
-	var line string
-	var readErr error
-	ch := make(chan struct{})
-	go func() {
-		rbuf := bufio.NewReader(os.Stdin)
-		line, readErr = rbuf.ReadString('\n')
-		close(ch)
-	}()
-	select {
-	case <-ctx.Done():
-		return "", ctx.Err()
-	case <-ch:
-		if readErr != nil {
-			return "", readErr
-		}
-		return strings.TrimRight(line, "\n"), nil
-	}
-}
-
-func (app *earthlyApp) actionOrgCreate(cliCtx *cli.Context) error {
-	app.commandName = "orgCreate"
-	if cliCtx.NArg() != 1 {
-		return errors.New("invalid number of arguments provided")
-	}
-	org := cliCtx.Args().Get(0)
-	cloudClient, err := cloud.NewClient(app.apiServer, app.sshAuthSock, app.authToken, app.console.Warnf)
-	if err != nil {
-		return errors.Wrap(err, "failed to create cloud client")
-	}
-	err = cloudClient.CreateOrg(cliCtx.Context, org)
-	if err != nil {
-		return errors.Wrap(err, "failed to create org")
-	}
-	return nil
-}
-
-func (app *earthlyApp) actionOrgList(cliCtx *cli.Context) error {
-	app.commandName = "orgList"
-	cloudClient, err := cloud.NewClient(app.apiServer, app.sshAuthSock, app.authToken, app.console.Warnf)
-	if err != nil {
-		return errors.Wrap(err, "failed to create cloud client")
-	}
-	orgs, err := cloudClient.ListOrgs(cliCtx.Context)
-	if err != nil {
-		return errors.Wrap(err, "failed to list orgs")
-	}
-
-	w := tabwriter.NewWriter(os.Stdout, 0, 0, 2, ' ', 0)
-	for _, org := range orgs {
-		fmt.Fprintf(w, "/%s/", org.Name)
-		if org.Admin {
-			fmt.Fprintf(w, "\tadmin")
-		} else {
-			fmt.Fprintf(w, "\tmember")
-		}
-		fmt.Fprintf(w, "\n")
-	}
-	w.Flush()
-
-	return nil
-}
-
-func (app *earthlyApp) actionOrgListPermissions(cliCtx *cli.Context) error {
-	app.commandName = "orgListPermissions"
-	if cliCtx.NArg() != 1 {
-		return errors.New("invalid number of arguments provided")
-	}
-	path := cliCtx.Args().Get(0)
-	if !strings.HasSuffix(path, "/") {
-		path += "/"
-	}
-	cloudClient, err := cloud.NewClient(app.apiServer, app.sshAuthSock, app.authToken, app.console.Warnf)
-	if err != nil {
-		return errors.Wrap(err, "failed to create cloud client")
-	}
-	orgs, err := cloudClient.ListOrgPermissions(cliCtx.Context, path)
-	if err != nil {
-		return errors.Wrap(err, "failed to list org permissions")
-	}
-
-	w := tabwriter.NewWriter(os.Stdout, 0, 0, 2, ' ', 0)
-	for _, org := range orgs {
-		fmt.Fprintf(w, "%s\t%s", org.Path, org.User)
-		if org.Write {
-			fmt.Fprintf(w, "\trw")
-		} else {
-			fmt.Fprintf(w, "\tr")
-		}
-		fmt.Fprintf(w, "\n")
-	}
-	w.Flush()
-	return nil
-}
-
-func (app *earthlyApp) actionOrgInvite(cliCtx *cli.Context) error {
-	app.commandName = "orgInvite"
-	if cliCtx.NArg() < 2 {
-		return errors.New("invalid number of arguments provided")
-	}
-	path := cliCtx.Args().Get(0)
-	if !strings.HasSuffix(path, "/") {
-		return errors.New("invitation paths must end with a slash (/)")
-	}
-
-	cloudClient, err := cloud.NewClient(app.apiServer, app.sshAuthSock, app.authToken, app.console.Warnf)
-	if err != nil {
-		return errors.Wrap(err, "failed to create cloud client")
-	}
-	userEmail := cliCtx.Args().Get(1)
-	err = cloudClient.Invite(cliCtx.Context, path, userEmail, app.writePermission)
-	if err != nil {
-		return errors.Wrap(err, "failed to invite user into org")
-	}
-	return nil
-}
-
-func (app *earthlyApp) actionOrgRevoke(cliCtx *cli.Context) error {
-	app.commandName = "orgRevoke"
-	if cliCtx.NArg() < 2 {
-		return errors.New("invalid number of arguments provided")
-	}
-	path := cliCtx.Args().Get(0)
-	if !strings.HasSuffix(path, "/") {
-		return errors.New("revoked paths must end with a slash (/)")
-	}
-
-	cloudClient, err := cloud.NewClient(app.apiServer, app.sshAuthSock, app.authToken, app.console.Warnf)
-	if err != nil {
-		return errors.Wrap(err, "failed to create cloud client")
-	}
-	userEmail := cliCtx.Args().Get(1)
-	err = cloudClient.RevokePermission(cliCtx.Context, path, userEmail)
-	if err != nil {
-		return errors.Wrap(err, "failed to revoke user from org")
-	}
-	return nil
-}
-
-func (app *earthlyApp) actionSecretsList(cliCtx *cli.Context) error {
-	app.commandName = "secretsList"
-
-	path := "/"
-	if cliCtx.NArg() > 1 {
-		return errors.New("invalid number of arguments provided")
-	} else if cliCtx.NArg() == 1 {
-		path = cliCtx.Args().Get(0)
-	}
-	if !strings.HasSuffix(path, "/") {
-		path += "/"
-	}
-	cloudClient, err := cloud.NewClient(app.apiServer, app.sshAuthSock, app.authToken, app.console.Warnf)
-	if err != nil {
-		return errors.Wrap(err, "failed to create cloud client")
-	}
-	paths, err := cloudClient.List(cliCtx.Context, path)
-	if err != nil {
-		return errors.Wrap(err, "failed to list secret")
-	}
-	for _, path := range paths {
-		fmt.Println(path)
-	}
-	return nil
-}
-
-func (app *earthlyApp) actionSecretsGet(cliCtx *cli.Context) error {
-	app.commandName = "secretsGet"
-	if cliCtx.NArg() != 1 {
-		return errors.New("invalid number of arguments provided")
-	}
-	path := cliCtx.Args().Get(0)
-	cloudClient, err := cloud.NewClient(app.apiServer, app.sshAuthSock, app.authToken, app.console.Warnf)
-	if err != nil {
-		return errors.Wrap(err, "failed to create cloud client")
-	}
-	data, err := cloudClient.Get(cliCtx.Context, path)
-	if err != nil {
-		return errors.Wrap(err, "failed to get secret")
-	}
-	fmt.Printf("%s", data)
-	if !app.disableNewLine {
-		fmt.Printf("\n")
-	}
-	return nil
-}
-
-func (app *earthlyApp) actionSecretsRemove(cliCtx *cli.Context) error {
-	app.commandName = "secretsRemove"
-	if cliCtx.NArg() != 1 {
-		return errors.New("invalid number of arguments provided")
-	}
-	path := cliCtx.Args().Get(0)
-	cloudClient, err := cloud.NewClient(app.apiServer, app.sshAuthSock, app.authToken, app.console.Warnf)
-	if err != nil {
-		return errors.Wrap(err, "failed to create cloud client")
-	}
-	err = cloudClient.Remove(cliCtx.Context, path)
-	if err != nil {
-		return errors.Wrap(err, "failed to remove secret")
-	}
-	return nil
-}
-
-func (app *earthlyApp) actionSecretsSet(cliCtx *cli.Context) error {
-	app.commandName = "secretsSet"
-	var path string
-	var value string
-	if app.secretFile == "" && !app.secretStdin {
-		if cliCtx.NArg() != 2 {
-			return errors.New("invalid number of arguments provided")
-		}
-		path = cliCtx.Args().Get(0)
-		value = cliCtx.Args().Get(1)
-	} else if app.secretStdin {
-		if app.secretFile != "" {
-			return errors.New("only one of --file or --stdin can be used at a time")
-		}
-		if cliCtx.NArg() != 1 {
-			return errors.New("invalid number of arguments provided")
-		}
-		path = cliCtx.Args().Get(0)
-		data, err := io.ReadAll(os.Stdin)
-		if err != nil {
-			return errors.Wrap(err, "failed to read from stdin")
-		}
-		value = string(data)
-	} else {
-		if cliCtx.NArg() != 1 {
-			return errors.New("invalid number of arguments provided")
-		}
-		path = cliCtx.Args().Get(0)
-		data, err := os.ReadFile(app.secretFile)
-		if err != nil {
-			return errors.Wrapf(err, "failed to read secret from %s", app.secretFile)
-		}
-		value = string(data)
-	}
-
-	cloudClient, err := cloud.NewClient(app.apiServer, app.sshAuthSock, app.authToken, app.console.Warnf)
-	if err != nil {
-		return errors.Wrap(err, "failed to create cloud client")
-	}
-	err = cloudClient.Set(cliCtx.Context, path, []byte(value))
-	if err != nil {
-		return errors.Wrap(err, "failed to set secret")
-	}
-	return nil
-}
-
-func (app *earthlyApp) actionRegister(cliCtx *cli.Context) error {
-	app.commandName = "secretsRegister"
-	if app.email == "" {
-		return errors.New("no email given")
-	}
-
-	if !strings.Contains(app.email, "@") {
-		return errors.New("email is invalid")
-	}
-
-	cloudClient, err := cloud.NewClient(app.apiServer, app.sshAuthSock, app.authToken, app.console.Warnf)
-	if err != nil {
-		return errors.Wrap(err, "failed to create cloud client")
-	}
-
-	if app.token == "" {
-		err := cloudClient.RegisterEmail(cliCtx.Context, app.email)
-		if err != nil {
-			return errors.Wrap(err, "failed to register email")
-		}
-		fmt.Printf("An email has been sent to %q containing a registration token\n", app.email)
-		return nil
-	}
-
-	var publicKeys []*agent.Key
-	if app.sshAuthSock != "" {
-		var err error
-		publicKeys, err = cloudClient.GetPublicKeys(cliCtx.Context)
-		if err != nil {
-			return err
-		}
-	}
-
-	// Our signal handling under main() doesn't cause reading from stdin to cancel
-	// as there's no way to pass app.ctx to stdin read calls.
-	signal.Reset(syscall.SIGINT, syscall.SIGTERM)
-
-	pword := app.password
-	if app.password == "" {
-		fmt.Printf("pick a password: ")
-		enteredPassword, err := term.ReadPassword(int(syscall.Stdin))
-		if err != nil {
-			return err
-		}
-		fmt.Println("")
-		fmt.Printf("confirm password: ")
-		enteredPassword2, err := term.ReadPassword(int(syscall.Stdin))
-		if err != nil {
-			return err
-		}
-		fmt.Println("")
-		if string(enteredPassword) != string(enteredPassword2) {
-			return errors.Errorf("passwords do not match")
-		}
-		pword = string(enteredPassword)
-	}
-
-	var interactiveAccept bool
-	if !app.termsConditionsPrivacy {
-		rawAccept, err := promptInput(cliCtx.Context, "I acknowledge Earthly Technologies’ Privacy Policy (https://earthly.dev/privacy-policy) and agree to Earthly Technologies Terms of Service (https://earthly.dev/tos) [y/N]: ")
-		if err != nil {
-			return err
-		}
-		if rawAccept == "" {
-			rawAccept = "n"
-		}
-		accept := strings.ToLower(rawAccept)[0]
-		interactiveAccept = (accept == 'y')
-	}
-	termsConditionsPrivacy := app.termsConditionsPrivacy || interactiveAccept
-
-	var publicKey string
-	if app.registrationPublicKey == "" {
-		if len(publicKeys) > 0 {
-			rawIsRegisterSSHKey, err := promptInput(cliCtx.Context, "Would you like to register an SSH key to be used as a form of login? [Y/n]: ")
-			if err != nil {
-				return err
-			}
-			if rawIsRegisterSSHKey == "" {
-				rawIsRegisterSSHKey = "y"
-			}
-			isRegisterSSHKey := (strings.ToLower(rawIsRegisterSSHKey)[0] == 'y')
-			if isRegisterSSHKey {
-				fmt.Printf("Which of the following keys do you want to register?\n")
-				fmt.Printf("0) none\n")
-				for i, key := range publicKeys {
-					fmt.Printf("%d) %s\n", i+1, key.String())
-				}
-				keyNum, err := promptInput(cliCtx.Context, "enter key number (1=default): ")
-				if err != nil {
-					return err
-				}
-				if keyNum == "" {
-					keyNum = "1"
-				}
-				i, err := strconv.Atoi(keyNum)
-				if err != nil {
-					return errors.Wrap(err, "invalid key number")
-				}
-				if i < 0 || i > len(publicKeys) {
-					return errors.Errorf("invalid key number")
-				}
-				if i > 0 {
-					publicKey = publicKeys[i-1].String()
-				}
-			}
-		}
-	} else {
-		_, _, _, _, err := ssh.ParseAuthorizedKey([]byte(app.registrationPublicKey))
-		if err == nil {
-			// supplied public key is valid
-			publicKey = app.registrationPublicKey
-		} else {
-			// otherwise see if it matches the name (Comment) of a key known by the ssh agent
-			for _, key := range publicKeys {
-				if key.Comment == app.registrationPublicKey {
-					publicKey = key.String()
-					break
-				}
-			}
-			if publicKey == "" {
-				return errors.Errorf("failed to find key in ssh agent's known keys")
-			}
-		}
-	}
-
-	err = cloudClient.CreateAccount(cliCtx.Context, app.email, app.token, pword, publicKey, termsConditionsPrivacy)
-	if err != nil {
-		return errors.Wrap(err, "failed to create account")
-	}
-
-	fmt.Println("Account registration complete")
-	return nil
-}
-
-func (app *earthlyApp) actionAccountListKeys(cliCtx *cli.Context) error {
-	app.commandName = "accountListKeys"
-	cloudClient, err := cloud.NewClient(app.apiServer, app.sshAuthSock, app.authToken, app.console.Warnf)
-	if err != nil {
-		return errors.Wrap(err, "failed to create cloud client")
-	}
-	keys, err := cloudClient.ListPublicKeys(cliCtx.Context)
-	if err != nil {
-		return errors.Wrap(err, "failed to list account keys")
-	}
-	for _, key := range keys {
-		fmt.Printf("%s\n", key)
-	}
-	return nil
-}
-
-func (app *earthlyApp) actionAccountAddKey(cliCtx *cli.Context) error {
-	app.commandName = "accountAddKey"
-	cloudClient, err := cloud.NewClient(app.apiServer, app.sshAuthSock, app.authToken, app.console.Warnf)
-	if err != nil {
-		return errors.Wrap(err, "failed to create cloud client")
-	}
-	if cliCtx.NArg() > 1 {
-		for _, k := range cliCtx.Args().Slice() {
-			err := cloudClient.AddPublickKey(cliCtx.Context, k)
-			if err != nil {
-				return errors.Wrap(err, "failed to add public key to account")
-			}
-		}
-		return nil
-	}
-
-	publicKeys, err := cloudClient.GetPublicKeys(cliCtx.Context)
-	if err != nil {
-		return err
-	}
-	if len(publicKeys) == 0 {
-		return errors.Errorf("unable to list available public keys, is ssh-agent running?; do you need to run ssh-add?")
-	}
-
-	// Our signal handling under main() doesn't cause reading from stdin to cancel
-	// as there's no way to pass app.ctx to stdin read calls.
-	signal.Reset(syscall.SIGINT, syscall.SIGTERM)
-
-	fmt.Printf("Which of the following keys do you want to register?\n")
-	for i, key := range publicKeys {
-		fmt.Printf("%d) %s\n", i+1, key.String())
-	}
-	keyNum, err := promptInput(cliCtx.Context, "enter key number (1=default): ")
-	if err != nil {
-		return err
-	}
-	if keyNum == "" {
-		keyNum = "1"
-	}
-	i, err := strconv.Atoi(keyNum)
-	if err != nil {
-		return errors.Wrap(err, "invalid key number")
-	}
-	if i <= 0 || i > len(publicKeys) {
-		return errors.Errorf("invalid key number")
-	}
-	publicKey := publicKeys[i-1].String()
-
-	err = cloudClient.AddPublickKey(cliCtx.Context, publicKey)
-	if err != nil {
-		return errors.Wrap(err, "failed to add public key to account")
-	}
-
-	// switch over to new key if the user is currently using password-based auth
-	email, authType, _, err := cloudClient.WhoAmI(cliCtx.Context)
-	if err != nil {
-		return errors.Wrap(err, "failed to validate auth token")
-	}
-	if authType == "password" {
-		err = cloudClient.SetSSHCredentials(cliCtx.Context, email, publicKey)
-		if err != nil {
-			app.console.Warnf("failed to authenticate using newly added public key: %s", err.Error())
-			return nil
-		}
-		fmt.Printf("Switching from password-based login to ssh-based login\n")
-	}
-
-	return nil
-}
-
-func (app *earthlyApp) actionAccountRemoveKey(cliCtx *cli.Context) error {
-	app.commandName = "accountRemoveKey"
-	cloudClient, err := cloud.NewClient(app.apiServer, app.sshAuthSock, app.authToken, app.console.Warnf)
-	if err != nil {
-		return errors.Wrap(err, "failed to create cloud client")
-	}
-	for _, k := range cliCtx.Args().Slice() {
-		err := cloudClient.RemovePublickKey(cliCtx.Context, k)
-		if err != nil {
-			return errors.Wrap(err, "failed to add public key to account")
-		}
-	}
-	return nil
-}
-func (app *earthlyApp) actionAccountListTokens(cliCtx *cli.Context) error {
-	app.commandName = "accountListTokens"
-	cloudClient, err := cloud.NewClient(app.apiServer, app.sshAuthSock, app.authToken, app.console.Warnf)
-	if err != nil {
-		return errors.Wrap(err, "failed to create cloud client")
-	}
-	tokens, err := cloudClient.ListTokens(cliCtx.Context)
-	if err != nil {
-		return errors.Wrap(err, "failed to list account tokens")
-	}
-	if len(tokens) == 0 {
-		return nil // avoid printing header columns when there are no tokens
-	}
-
-	now := time.Now()
-
-	w := tabwriter.NewWriter(os.Stdout, 0, 0, 2, ' ', 0)
-	fmt.Fprintf(w, "Token Name\tRead/Write\tExpiry\n")
-	for _, token := range tokens {
-		expired := now.After(token.Expiry)
-		fmt.Fprintf(w, "%s", token.Name)
-		if token.Write {
-			fmt.Fprintf(w, "\trw")
-		} else {
-			fmt.Fprintf(w, "\tr")
-		}
-		fmt.Fprintf(w, "\t%s UTC", token.Expiry.UTC().Format("2006-01-02T15:04"))
-		if expired {
-			fmt.Fprintf(w, " *expired*")
-		}
-		fmt.Fprintf(w, "\n")
-	}
-	w.Flush()
-	return nil
-}
-func (app *earthlyApp) actionAccountCreateToken(cliCtx *cli.Context) error {
-	app.commandName = "accountCreateToken"
-	if cliCtx.NArg() != 1 {
-		return errors.New("invalid number of arguments provided")
-	}
-
-	var expiry time.Time
-	if app.expiry == "" {
-		expiry = time.Now().Add(time.Hour * 24 * 365)
-	} else if app.expiry == "never" {
-		expiry = time.Now().Add(time.Hour * 24 * 365 * 100) // TODO save this some other way
-	} else {
-		layouts := []string{
-			"2006-01-02",
-			time.RFC3339,
-		}
-
-		var err error
-		for _, layout := range layouts {
-			expiry, err = time.Parse(layout, app.expiry)
-			if err == nil {
-				break
-			}
-		}
-		if err != nil {
-			return errors.Errorf("failed to parse expiry %q", app.expiry)
-		}
-	}
-
-	cloudClient, err := cloud.NewClient(app.apiServer, app.sshAuthSock, app.authToken, app.console.Warnf)
-	if err != nil {
-		return errors.Wrap(err, "failed to create cloud client")
-	}
-	name := cliCtx.Args().First()
-	token, err := cloudClient.CreateToken(cliCtx.Context, name, app.writePermission, &expiry)
-	if err != nil {
-		return errors.Wrap(err, "failed to create token")
-	}
-	expiryStr := humanize.Time(expiry)
-	fmt.Printf("created token %q which will expire in %s; save this token somewhere, it can't be viewed again (only reset)\n", token, expiryStr)
-	return nil
-}
-func (app *earthlyApp) actionAccountRemoveToken(cliCtx *cli.Context) error {
-	app.commandName = "accountRemoveToken"
-	if cliCtx.NArg() != 1 {
-		return errors.New("invalid number of arguments provided")
-	}
-	name := cliCtx.Args().First()
-	cloudClient, err := cloud.NewClient(app.apiServer, app.sshAuthSock, app.authToken, app.console.Warnf)
-	if err != nil {
-		return errors.Wrap(err, "failed to create cloud client")
-	}
-	err = cloudClient.RemoveToken(cliCtx.Context, name)
-	if err != nil {
-		return errors.Wrap(err, "failed to remove account tokens")
-	}
-	return nil
-}
-
-func (app *earthlyApp) actionAccountLogin(cliCtx *cli.Context) error {
-	app.commandName = "accountLogin"
-	email := app.email
-	token := app.token
-	pass := app.password
-
-	if cliCtx.NArg() == 1 {
-		emailOrToken := cliCtx.Args().First()
-		if token == "" && email == "" {
-			if cloud.IsValidEmail(cliCtx.Context, emailOrToken) {
-				email = emailOrToken
-			} else {
-				token = emailOrToken
-			}
-
-		} else {
-			return errors.New("invalid number of arguments provided")
-		}
-	} else if cliCtx.NArg() > 1 {
-		return errors.New("invalid number of arguments provided")
-	}
-
-	if token != "" && (email != "" || pass != "") {
-		return errors.New("--token cannot be used in conjuction with --email or --password")
-	}
-	cloudClient, err := cloud.NewClient(app.apiServer, app.sshAuthSock, app.authToken, app.console.Warnf)
-	if err != nil {
-		return errors.Wrap(err, "failed to create cloud client")
-	}
-
-	// special case where global auth token overrides login logic
-	if app.authToken != "" {
-		if email != "" || token != "" || pass != "" {
-			return errLoginFlagsHaveNoEffect
-		}
-		loggedInEmail, authType, writeAccess, err := cloudClient.WhoAmI(cliCtx.Context)
-		if err != nil {
-			return errors.Wrap(err, "failed to validate auth token")
-		}
-		if !writeAccess {
-			authType = "read-only-" + authType
-		}
-		app.console.Printf("Logged in as %q using %s auth\n", loggedInEmail, authType)
-		app.printLogSharingMessage()
-		return nil
-	}
-
-	err = cloudClient.DeleteCachedToken(cliCtx.Context)
-	if err != nil {
-		return err
-	}
-
-	if token != "" || pass != "" {
-		err := cloudClient.DeleteAuthCache(cliCtx.Context)
-		if err != nil {
-			return errors.Wrap(err, "failed to clear cached credentials")
-		}
-		cloudClient.DisableSSHKeyGuessing(cliCtx.Context)
-	} else if email != "" {
-		if err = cloudClient.FindSSHCredentials(cliCtx.Context, email); err == nil {
-			// if err is not nil, we will try again below via cloudClient.WhoAmI()
-			err = cloudClient.Authenticate(cliCtx.Context)
-			if err != nil {
-				return errors.Wrap(err, "authentication with cloud server failed")
-			}
-			app.console.Printf("Logged in as %q using ssh auth\n", email)
-			app.printLogSharingMessage()
-			return nil
-		}
-	}
-
-	loggedInEmail, authType, writeAccess, err := cloudClient.WhoAmI(cliCtx.Context)
-	switch errors.Cause(err) {
-	case cloud.ErrUnauthorized:
-		break
-	case nil:
-		if email != "" && email != loggedInEmail {
-			break // case where a user has multiple emails and wants to switch
-		}
-		if !writeAccess {
-			authType = "read-only-" + authType
-		}
-		app.console.Printf("Logged in as %q using %s auth\n", loggedInEmail, authType)
-		app.printLogSharingMessage()
-		return nil
-	default:
-		return err
-	}
-
-	if email == "" && token == "" {
-		if app.sshAuthSock == "" {
-			app.console.Warnf("No ssh auth socket detected; falling back to password-based login\n")
-		}
-
-		emailOrToken, err := promptInput(cliCtx.Context, "enter your email or auth token: ")
-		if err != nil {
-			return err
-		}
-		if strings.Contains(emailOrToken, "@") {
-			email = emailOrToken
-		} else {
-			token = emailOrToken
-		}
-	}
-
-	if email != "" && pass == "" {
-		app.console.Printf("enter your password: \n")
-		passwordBytes, err := term.ReadPassword(int(syscall.Stdin))
-		if err != nil {
-			return err
-		}
-		pass = string(passwordBytes)
-		if pass == "" {
-			return errors.Errorf("no password entered")
-		}
-	}
-
-	if token != "" {
-		email, err = cloudClient.SetTokenCredentials(cliCtx.Context, token)
-		if err != nil {
-			return err
-		}
-		app.console.Printf("Logged in as %q using token auth\n", email) // TODO display if using read-only token
-		app.printLogSharingMessage()
-	} else {
-		err = cloudClient.SetPasswordCredentials(cliCtx.Context, email, string(pass))
-		if err != nil {
-			return err
-		}
-		app.console.Printf("Logged in as %q using password auth\n", email)
-		app.console.Printf("Warning unencrypted password has been stored under ~/.earthly/auth.credentials; consider using ssh-based auth to prevent this.\n")
-		app.printLogSharingMessage()
-	}
-	err = cloudClient.Authenticate(cliCtx.Context)
-	if err != nil {
-		return errors.Wrap(err, "authentication with cloud server failed")
-	}
-	return nil
-}
-
-func (app *earthlyApp) printLogSharingMessage() {
-	app.console.Printf("Log sharing is enabled by default. If you would like to disable it, run:\n" +
-		"\n" +
-		"\tearthly config global.disable_log_sharing true")
-}
-
-func (app *earthlyApp) actionAccountLogout(cliCtx *cli.Context) error {
-	app.commandName = "accountLogout"
-
-	if app.authToken != "" {
-		return errLogoutHasNoEffectWhenAuthTokenSet
-	}
-
-	cloudClient, err := cloud.NewClient(app.apiServer, app.sshAuthSock, app.authToken, app.console.Warnf)
-	if err != nil {
-		return err
-	}
-	err = cloudClient.DeleteAuthCache(cliCtx.Context)
-	if err != nil {
-		return errors.Wrap(err, "failed to logout")
-	}
-	return nil
-}
-
-func (app *earthlyApp) actionDebugAst(cliCtx *cli.Context) error {
-	app.commandName = "debugAst"
-	if cliCtx.NArg() > 1 {
-		return errors.New("invalid number of arguments provided")
-	}
-	path := "./Earthfile"
-	if cliCtx.NArg() == 1 {
-		path = cliCtx.Args().First()
-	}
-
-	ef, err := ast.Parse(cliCtx.Context, path, app.enableSourceMap)
-	if err != nil {
-		return err
-	}
-	efDt, err := json.Marshal(ef)
-	if err != nil {
-		return errors.Wrap(err, "marshal ast")
-	}
-	fmt.Print(string(efDt))
-	return nil
-}
-
-func (app *earthlyApp) actionPrune(cliCtx *cli.Context) error {
-	app.commandName = "prune"
-	if cliCtx.NArg() != 0 {
-		return errors.New("invalid arguments")
-	}
-	if app.pruneReset {
-		err := buildkitd.ResetCache(cliCtx.Context, app.console, app.buildkitdImage, app.containerName, app.containerFrontend, app.buildkitdSettings)
-		if err != nil {
-			return errors.Wrap(err, "reset cache")
-		}
-		return nil
-	}
-
-	if app.isUsingSatellite(cliCtx) {
-		return errors.New("Cannot prune when using a satellite")
-	}
-
-	// Prune via API.
-	bkClient, err := app.GetBuildkitClient(cliCtx, nil)
-	if err != nil {
-		return errors.Wrap(err, "prune new buildkitd client")
-	}
-	defer bkClient.Close()
-	var opts []client.PruneOption
-	if app.pruneAll {
-		opts = append(opts, client.PruneAll)
-	}
-	ch := make(chan client.UsageInfo, 1)
-	eg, ctx := errgroup.WithContext(cliCtx.Context)
-	eg.Go(func() error {
-		err = bkClient.Prune(ctx, ch, opts...)
-		if err != nil {
-			return errors.Wrap(err, "buildkit prune")
-		}
-		close(ch)
-		return nil
-	})
-	eg.Go(func() error {
-		for {
-			select {
-			case _, ok := <-ch:
-				if !ok {
-					return nil
-				}
-				// TODO: Print some progress info.
-			case <-ctx.Done():
-				return nil
-			}
-		}
-	})
-	err = eg.Wait()
-	if err != nil {
-		return errors.Wrap(err, "err group")
-	}
-	return nil
-}
-
-func (app *earthlyApp) actionDocker(cliCtx *cli.Context) error {
-	app.commandName = "docker"
-
-	dir := filepath.Dir(app.dockerfilePath)
-	earthfilePath := filepath.Join(dir, "Earthfile")
-	earthfilePathExists, err := fileutil.FileExists(earthfilePath)
-	if err != nil {
-		return errors.Wrapf(err, "failed to check if %s exists", earthfilePath)
-	}
-	if earthfilePathExists {
-		return errors.Errorf("earthfile already exists; please delete it if you wish to continue")
-	}
-	defer os.Remove(earthfilePath)
-
-	err = docker2earthly.Docker2Earthly(app.dockerfilePath, earthfilePath, app.earthfileFinalImage)
-	if err != nil {
-		return err
-	}
-
-	fmt.Fprintf(os.Stderr, "Warning: earthly does not support all dockerfile commands and is highly experimental as a result, use with caution.\n")
-
-	app.imageMode = false
-	app.artifactMode = false
-	app.interactiveDebugging = true
-	flagArgs := []string{}
-	nonFlagArgs := []string{"+build"}
-
-	return app.actionBuildImp(cliCtx, flagArgs, nonFlagArgs)
-}
-
-func (app *earthlyApp) actionDocker2Earthly(cliCtx *cli.Context) error {
-	app.commandName = "docker2earthly"
-	err := docker2earthly.Docker2Earthly(app.dockerfilePath, app.earthfilePath, app.earthfileFinalImage)
-	if err != nil {
-		return err
-	}
-	fmt.Fprintf(os.Stderr, "An Earthfile has been generated; to run it use: earthly +build; then run with docker run -ti %s\n", app.earthfileFinalImage)
-	return nil
-}
-
-func (app *earthlyApp) actionConfig(cliCtx *cli.Context) error {
-	app.commandName = "config"
-	if cliCtx.NArg() != 2 {
-		return errors.New("invalid number of arguments provided")
-	}
-
-	args := cliCtx.Args().Slice()
-	inConfig, err := config.ReadConfigFile(app.configPath)
-	if err != nil {
-		if cliCtx.IsSet("config") || !errors.Is(err, os.ErrNotExist) {
-			return errors.Wrapf(err, "read config")
-		}
-	}
-
-	var outConfig []byte
-
-	switch args[1] {
-	case "-h", "--help":
-		if err = config.PrintHelp(args[0]); err != nil {
-			return errors.Wrap(err, "help")
-		}
-		return nil // exit now without writing any changes to config
-	case "--delete":
-		outConfig, err = config.Delete(inConfig, args[0])
-		if err != nil {
-			return errors.Wrap(err, "delete config")
-		}
-	default:
-		// args are key/value pairs, e.g. ["global.conversion_parallelism","5"]
-		outConfig, err = config.Upsert(inConfig, args[0], args[1])
-		if err != nil {
-			return errors.Wrap(err, "upsert config")
-		}
-	}
-
-	if app.configDryRun {
-		fmt.Println(string(outConfig))
-		return nil
-	}
-
-	err = config.WriteConfigFile(app.configPath, outConfig)
-	if err != nil {
-		return errors.Wrap(err, "write config")
-	}
-
-	return nil
-}
-
-func (app *earthlyApp) actionBuild(cliCtx *cli.Context) error {
-	app.commandName = "build"
-
-	if app.ci {
-		app.useInlineCache = true
-		app.noOutput = !app.output && !app.artifactMode && !app.imageMode
-		app.strict = true
-		if app.remoteCache == "" && app.push {
-			app.saveInlineCache = true
-		}
-
-		if app.interactiveDebugging {
-			return errors.New("unable to use --ci flag in combination with --interactive flag")
-		}
-	}
-	if app.interactiveDebugging {
-		if !termutil.IsTTY() {
-			return errors.New("A tty-terminal must be present in order to the --interactive flag")
-		}
-		if !containerutil.IsLocal(app.buildkitHost) {
-			return errors.New("the --interactive flag is not currently supported with non-local buildkit servers")
-		}
-	}
-
-	if app.imageMode && app.artifactMode {
-		return errors.New("both image and artifact modes cannot be active at the same time")
-	}
-	if (app.imageMode && app.noOutput) || (app.artifactMode && app.noOutput) {
-		if app.ci {
-			app.noOutput = false
-		} else {
-			return errors.New("cannot use --no-output with image or artifact modes")
-		}
-	}
-
-	flagArgs, nonFlagArgs, err := variables.ParseFlagArgsWithNonFlags(cliCtx.Args().Slice())
-	if err != nil {
-		return errors.Wrapf(err, "parse args %s", strings.Join(cliCtx.Args().Slice(), " "))
-	}
-
-	return app.actionBuildImp(cliCtx, flagArgs, nonFlagArgs)
-}
-
-// warnIfArgContainsBuildArg will issue a warning if a flag is incorrectly prefixed with build-arg.
-// TODO this check should be replaced with a warning if an arg was given but never used.
-func (app *earthlyApp) warnIfArgContainsBuildArg(flagArgs []string) {
-	for _, flag := range flagArgs {
-		if strings.HasPrefix(flag, "build-arg=") || strings.HasPrefix(flag, "buildarg=") {
-			app.console.Warnf("Found a flag named %q; flags after the build target should be specified as --KEY=VAL\n", flag)
-		}
-	}
-}
-
-func (app *earthlyApp) combineVariables(dotEnvMap map[string]string, flagArgs []string) (*variables.Scope, error) {
-	dotEnvVars := variables.NewScope()
-	for k, v := range dotEnvMap {
-		dotEnvVars.AddInactive(k, v)
-	}
-	buildArgs := append([]string{}, app.buildArgs.Value()...)
-	buildArgs = append(buildArgs, flagArgs...)
-	overridingVars, err := variables.ParseCommandLineArgs(buildArgs)
-	if err != nil {
-		return nil, errors.Wrap(err, "parse build args")
-	}
-	return variables.CombineScopes(overridingVars, dotEnvVars), nil
-}
-
-func (app *earthlyApp) actionBuildImp(cliCtx *cli.Context, flagArgs, nonFlagArgs []string) error {
-	var target domain.Target
-	var artifact domain.Artifact
-	destPath := "./"
-	if app.imageMode {
-		if len(nonFlagArgs) == 0 {
-			cli.ShowAppHelp(cliCtx)
-			return errors.Errorf(
-				"no image reference provided. Try %s --image +<target-name>", cliCtx.App.Name)
-		} else if len(nonFlagArgs) != 1 {
-			cli.ShowAppHelp(cliCtx)
-			return errors.Errorf("invalid arguments %s", strings.Join(nonFlagArgs, " "))
-		}
-		targetName := nonFlagArgs[0]
-		var err error
-		target, err = domain.ParseTarget(targetName)
-		if err != nil {
-			return errors.Wrapf(err, "parse target name %s", targetName)
-		}
-	} else if app.artifactMode {
-		if len(nonFlagArgs) == 0 {
-			cli.ShowAppHelp(cliCtx)
-			return errors.Errorf(
-				"no artifact reference provided. Try %s --artifact +<target-name>/<artifact-name>", cliCtx.App.Name)
-		} else if len(nonFlagArgs) > 2 {
-			cli.ShowAppHelp(cliCtx)
-			return errors.Errorf("invalid arguments %s", strings.Join(nonFlagArgs, " "))
-		}
-		artifactName := nonFlagArgs[0]
-		if len(nonFlagArgs) == 2 {
-			destPath = nonFlagArgs[1]
-		}
-		var err error
-		artifact, err = domain.ParseArtifact(artifactName)
-		if err != nil {
-			return errors.Wrapf(err, "parse artifact name %s", artifactName)
-		}
-		target = artifact.Target
-	} else {
-		if len(nonFlagArgs) == 0 {
-			cli.ShowAppHelp(cliCtx)
-			return errors.Errorf(
-				"no target reference provided. Try %s +<target-name>", cliCtx.App.Name)
-		} else if len(nonFlagArgs) != 1 {
-			cli.ShowAppHelp(cliCtx)
-			return errors.Errorf("invalid arguments %s", strings.Join(nonFlagArgs, " "))
-		}
-		targetName := nonFlagArgs[0]
-		var err error
-		target, err = domain.ParseTarget(targetName)
-		if err != nil {
-			return errors.Wrapf(err, "parse target name %s", targetName)
-		}
-	}
-
-	cleanCollection := cleanup.NewCollection()
-	defer cleanCollection.Close()
-
-	cloudClient, err := cloud.NewClient(app.apiServer, app.sshAuthSock, app.authToken, app.console.Warnf)
-	if err != nil {
-		return errors.Wrap(err, "failed to create cloud client")
-	}
-
-	// Default upload logs, unless explicitly configured
-	if !app.cfg.Global.DisableLogSharing {
-		if cloudClient.IsLoggedIn(cliCtx.Context) {
-			// If you are logged in, then add the bundle builder code, and configure cleanup and post-build messages.
-			app.console = app.console.WithLogBundleWriter(target.String(), cleanCollection)
-
-			defer func() { // Defer this to keep log upload code together
-				logPath, err := app.console.WriteBundleToDisk()
-				if err != nil {
-					err := errors.Wrapf(err, "failed to write log to disk")
-					app.console.Warnf(err.Error())
-					return
-				}
-
-				id, err := cloudClient.UploadLog(cliCtx.Context, logPath)
-				if err != nil {
-					err := errors.Wrapf(err, "failed to upload log")
-					app.console.Warnf(err.Error())
-					return
-				}
-				app.console.Printf("Shareable link: %s\n", id)
-			}()
-		} else {
-			defer func() { // Defer this to keep log upload code together
-				app.console.Printf("Share your logs with an Earthly account (experimental)! Register for one at https://ci.earthly.dev.")
-			}()
-		}
-	}
-
-	app.console.PrintPhaseHeader(builder.PhaseInit, false, "")
-	app.warnIfArgContainsBuildArg(flagArgs)
-
-	bkClient, err := app.GetBuildkitClient(cliCtx, cloudClient)
-	if err != nil {
-		return errors.Wrap(err, "build new buildkitd client")
-	}
-	defer bkClient.Close()
-	isLocal := containerutil.IsLocal(app.buildkitdSettings.BuildkitAddress)
-	app.analyticsMetadata.isRemoteBuildkit = !isLocal
-
-	bkIP, err := buildkitd.GetContainerIP(cliCtx.Context, app.containerName, app.containerFrontend, app.buildkitdSettings)
-	if err != nil {
-		return errors.Wrap(err, "get buildkit container IP")
-	}
-
-	nativePlatform, err := platutil.GetNativePlatformViaBkClient(cliCtx.Context, bkClient)
-	if err != nil {
-		return errors.Wrap(err, "get native platform via buildkit client")
-	}
-	platr := platutil.NewResolver(nativePlatform)
-	app.analyticsMetadata.buildkitPlatform = platforms.Format(nativePlatform)
-	app.analyticsMetadata.userPlatform = platforms.Format(platr.LLBUser())
-	platr.AllowNativeAndUser = true
-	platformsSlice := make([]platutil.Platform, 0, len(app.platformsStr.Value()))
-	for _, p := range app.platformsStr.Value() {
-		platform, err := platr.Parse(p)
-		if err != nil {
-			return errors.Wrapf(err, "parse platform %s", p)
-		}
-		platformsSlice = append(platformsSlice, platform)
-	}
-	switch len(platformsSlice) {
-	case 0:
-	case 1:
-		platr.UpdatePlatform(platformsSlice[0])
-	default:
-		return errors.Errorf("multi-platform builds are not yet supported on the command line. You may, however, create a target with the instruction BUILD --plaform ... --platform ... %s", target)
-	}
-
-	dotEnvMap, err := godotenv.Read(app.envFile)
-	if err != nil {
-		// ignore ErrNotExist when using default .env file
-		if app.envFile != defaultEnvFile || !errors.Is(err, os.ErrNotExist) {
-			return errors.Wrapf(err, "read %s", app.envFile)
-		}
-	}
-
-	secretsMap, err := processSecrets(app.secrets.Value(), app.secretFiles.Value(), dotEnvMap)
-	if err != nil {
-		return err
-	}
-
-	debuggerSettings := debuggercommon.DebuggerSettings{
-		DebugLevelLogging: app.debug,
-		Enabled:           app.interactiveDebugging,
-		RepeaterAddr:      fmt.Sprintf("%s:8373", bkIP),
-		Term:              os.Getenv("TERM"),
-	}
-	if app.interactiveDebugging {
-		analytics.Count("features", "interactive-debugging")
-	}
-
-	debuggerSettingsData, err := json.Marshal(&debuggerSettings)
-	if err != nil {
-		return errors.Wrap(err, "debugger settings json marshal")
-	}
-	secretsMap[debuggercommon.DebuggerSettingsSecretsKey] = debuggerSettingsData
-
-	localhostProvider, err := localhostprovider.NewLocalhostProvider()
-	if err != nil {
-		return errors.Wrap(err, "failed to create localhostprovider")
-	}
-
-	cacheLocalDir, err := os.MkdirTemp("", "earthly-cache")
-	if err != nil {
-		return errors.Wrap(err, "make temp dir for cache")
-	}
-	defer os.RemoveAll(cacheLocalDir)
-	defaultLocalDirs := make(map[string]string)
-	defaultLocalDirs["earthly-cache"] = cacheLocalDir
-	buildContextProvider := provider.NewBuildContextProvider(app.console)
-	buildContextProvider.AddDirs(defaultLocalDirs)
-
-	customSecretProviderCmd, err := secretprovider.NewSecretProviderCmd(app.cfg.Global.SecretProvider)
-	if err != nil {
-		return errors.Wrap(err, "NewSecretProviderCmd")
-	}
-	secretProvider := secretprovider.New(
-		customSecretProviderCmd,
-		secretprovider.NewMapStore(secretsMap),
-		secretprovider.NewCloudStore(cloudClient),
-	)
-
-	attachables := []session.Attachable{
-		secretProvider,
-		buildContextProvider,
-		localhostProvider,
-	}
-
-	switch app.containerFrontend.Config().Setting {
-	case containerutil.FrontendDocker, containerutil.FrontendDockerShell:
-		attachables = append(attachables, authprovider.NewDockerAuthProvider(os.Stderr))
-
-	case containerutil.FrontendPodman, containerutil.FrontendPodmanShell:
-		attachables = append(attachables, authprovider.NewPodmanAuthProvider(os.Stderr))
-
-	default:
-		// Old default behavior
-		attachables = append(attachables, authprovider.NewDockerAuthProvider(os.Stderr))
-	}
-
-	gitLookup := buildcontext.NewGitLookup(app.console, app.sshAuthSock)
-	err = app.updateGitLookupConfig(gitLookup)
-	if err != nil {
-		return err
-	}
-
-	if app.sshAuthSock != "" {
-		ssh, err := sshprovider.NewSSHAgentProvider([]sshprovider.AgentConfig{{
-			Paths: []string{app.sshAuthSock},
-		}})
-		if err != nil {
-			return errors.Wrap(err, "ssh agent provider")
-		}
-		attachables = append(attachables, ssh)
-	}
-
-	var enttlmnts []entitlements.Entitlement
-	if app.allowPrivileged {
-		enttlmnts = append(enttlmnts, entitlements.EntitlementSecurityInsecure)
-	}
-
-	if termutil.IsTTY() {
-		go func() {
-			// Dialing does not accept URLs, it accepts an address and a "network". These cannot be handled as URL schemes.
-			// Since Shellrepeater hard-codes TCP, we drop it here and log the error if we fail to connect.
-
-			u, err := url.Parse(app.debuggerHost)
-			if err != nil {
-				panic("debugger host was not a URL")
-			}
-
-			debugTermConsole := app.console.WithPrefix("internal-term")
-			err = terminal.ConnectTerm(cliCtx.Context, u.Host, debugTermConsole)
-			if err != nil {
-				debugTermConsole.VerbosePrintf("unable to connect to terminal: %s", err.Error())
-			}
-		}()
-	}
-
-	overridingVars, err := app.combineVariables(dotEnvMap, flagArgs)
-	if err != nil {
-		return err
-	}
-
-	imageResolveMode := llb.ResolveModePreferLocal
-	if app.pull {
-		imageResolveMode = llb.ResolveModeForcePull
-	}
-
-	cacheImports := make(map[string]bool)
-	if app.remoteCache != "" {
-		cacheImports[app.remoteCache] = true
-	}
-	var cacheExport string
-	var maxCacheExport string
-	if app.remoteCache != "" && app.push {
-		if app.maxRemoteCache {
-			maxCacheExport = app.remoteCache
-		} else {
-			cacheExport = app.remoteCache
-		}
-	}
-	var parallelism semutil.Semaphore
-	if app.cfg.Global.ConversionParallelism != 0 {
-		parallelism = semutil.NewWeighted(int64(app.cfg.Global.ConversionParallelism))
-	}
-	localRegistryAddr := ""
-	if isLocal && app.localRegistryHost != "" {
-		lrURL, err := url.Parse(app.localRegistryHost)
-		if err != nil {
-			return errors.Wrapf(err, "parse local registry host %s", app.localRegistryHost)
-		}
-		localRegistryAddr = lrURL.Host
-	}
-	builderOpts := builder.Opt{
-		BkClient:               bkClient,
-		Console:                app.console,
-		Verbose:                app.verbose,
-		Attachables:            attachables,
-		Enttlmnts:              enttlmnts,
-		NoCache:                app.noCache,
-		CacheImports:           states.NewCacheImports(cacheImports),
-		CacheExport:            cacheExport,
-		MaxCacheExport:         maxCacheExport,
-		UseInlineCache:         app.useInlineCache,
-		SaveInlineCache:        app.saveInlineCache,
-		SessionID:              app.sessionID,
-		ImageResolveMode:       imageResolveMode,
-		CleanCollection:        cleanCollection,
-		OverridingVars:         overridingVars,
-		BuildContextProvider:   buildContextProvider,
-		GitLookup:              gitLookup,
-		UseFakeDep:             !app.noFakeDep,
-		Strict:                 app.strict,
-		DisableNoOutputUpdates: app.interactiveDebugging,
-		ParallelConversion:     (app.cfg.Global.ConversionParallelism != 0),
-		Parallelism:            parallelism,
-		LocalRegistryAddr:      localRegistryAddr,
-		FeatureFlagOverrides:   app.featureFlagOverrides,
-		ContainerFrontend:      app.containerFrontend,
-	}
-	b, err := builder.NewBuilder(cliCtx.Context, builderOpts)
-	if err != nil {
-		return errors.Wrap(err, "new builder")
-	}
-
-	app.console.PrintPhaseFooter(builder.PhaseInit, false, "")
-
-	builtinArgs := variables.DefaultArgs{
-		EarthlyVersion:  Version,
-		EarthlyBuildSha: GitSha,
-	}
-	buildOpts := builder.BuildOpt{
-		PrintPhases:                true,
-		Push:                       app.push,
-		NoOutput:                   app.noOutput,
-		OnlyFinalTargetImages:      app.imageMode,
-		PlatformResolver:           platr,
-		EnableGatewayClientLogging: app.debug,
-		BuiltinArgs:                builtinArgs,
-
-		// explicitly set this to true at the top level (without granting the entitlements.EntitlementSecurityInsecure buildkit option),
-		// to differentiate between a user forgetting to run earthly -P, versus a remotely referening an earthfile that requires privileged.
-		AllowPrivileged: true,
-	}
-	if app.artifactMode {
-		buildOpts.OnlyArtifact = &artifact
-		buildOpts.OnlyArtifactDestPath = destPath
-	}
-	_, err = b.BuildTarget(cliCtx.Context, target, buildOpts)
-	if err != nil {
-		return errors.Wrap(err, "build target")
-	}
-
-	return nil
-}
-
-func (app *earthlyApp) updateGitLookupConfig(gitLookup *buildcontext.GitLookup) error {
-	for k, v := range app.cfg.Git {
-		if k == "github" || k == "gitlab" || k == "bitbucket" {
-			app.console.Warnf("git configuration for %q found, did you mean %q?\n", k, k+".com")
-		}
-		pattern := v.Pattern
-		if pattern == "" {
-			// if empty, assume it will be of the form host.com/user/repo.git
-			host := k
-			if !strings.Contains(host, ".") {
-				host += ".com"
-			}
-			pattern = host + "/[^/]+/[^/]+"
-		}
-		auth := v.Auth
-		suffix := v.Suffix
-		if suffix == "" {
-			suffix = ".git"
-		}
-		err := gitLookup.AddMatcher(k, pattern, v.Substitute, v.User, v.Password, v.Prefix, suffix, auth, v.ServerKey, ifNilBoolDefault(v.StrictHostKeyChecking, true), v.Port)
-		if err != nil {
-			return errors.Wrap(err, "gitlookup")
-		}
-	}
-	return nil
-}
-
-func ifNilBoolDefault(ptr *bool, defaultValue bool) bool {
-	if ptr == nil {
-		return defaultValue
-	}
-	return *ptr
-}
-
-func (app *earthlyApp) actionListTargets(cliCtx *cli.Context) error {
-	app.commandName = "listTargets"
-
-	if cliCtx.NArg() > 1 {
-		return errors.New("invalid number of arguments provided")
-	}
-	var targetToParse string
-	if cliCtx.NArg() > 0 {
-		targetToParse = cliCtx.Args().Get(0)
-		if !(strings.HasPrefix(targetToParse, "/") || strings.HasPrefix(targetToParse, ".")) {
-			return errors.New("remote-paths are not currently supported; local paths must start with \"/\" or \".\"")
-		}
-		if strings.Contains(targetToParse, "+") {
-			return errors.New("path cannot contain a +")
-		}
-		targetToParse = strings.TrimSuffix(targetToParse, "/Earthfile")
-	}
-
-	targetToDisplay := targetToParse
-	if targetToParse == "" {
-		targetToDisplay = "current directory"
-	}
-
-	gitLookup := buildcontext.NewGitLookup(app.console, app.sshAuthSock)
-	resolver := buildcontext.NewResolver("", nil, gitLookup, app.console, "")
-	var gwClient gwclient.Client // TODO this is a nil pointer which causes a panic if we try to expand a remotely referenced earthfile
-	// it's expensive to create this gwclient, so we need to implement a lazy eval which returns it when required.
-
-	target, err := domain.ParseTarget(fmt.Sprintf("%s+base", targetToParse)) // the +base is required to make ParseTarget work; however is ignored by GetTargets
-	if err != nil {
-		return errors.Errorf("unable to locate Earthfile under %s", targetToDisplay)
-	}
-
-	targets, err := earthfile2llb.GetTargets(cliCtx.Context, resolver, gwClient, target)
-	if err != nil {
-		return errors.Errorf("unable to locate Earthfile under %s", targetToDisplay)
-	}
-	targets = append(targets, "base")
-	sort.Strings(targets)
-	for _, t := range targets {
-		var args []string
-		if t != "base" {
-			target.Target = t
-			args, err = earthfile2llb.GetTargetArgs(cliCtx.Context, resolver, gwClient, target)
-			if err != nil {
-				return err
-			}
-		}
-		if app.lsShowLong {
-			fmt.Printf("%s+%s\n", targetToParse, t)
-		} else {
-			fmt.Printf("+%s\n", t)
-		}
-		if app.lsShowArgs {
-			for _, arg := range args {
-				fmt.Printf("  --%s\n", arg)
-			}
-		}
-	}
-	return nil
-}
-
-func (app *earthlyApp) useSatellite(cliCtx *cli.Context, satelliteName, orgName string) error {
-	inConfig, err := config.ReadConfigFile(app.configPath)
-	if err != nil {
-		if cliCtx.IsSet("config") || !errors.Is(err, os.ErrNotExist) {
-			return errors.Wrap(err, "read config")
-		}
-	}
-
-	newConfig, err := config.Upsert(inConfig, "satellite.name", satelliteName)
-	if err != nil {
-		return errors.Wrap(err, "could not update satellite name")
-	}
-	// Update in-place so we can use it later, assuming the config change was successful.
-	app.cfg.Satellite.Name = satelliteName
-
-	newConfig, err = config.Upsert(newConfig, "satellite.org", orgName)
-	if err != nil {
-		return errors.Wrap(err, "could not update satellite name")
-	}
-	app.cfg.Satellite.Org = orgName
-	err = config.WriteConfigFile(app.configPath, newConfig)
-	if err != nil {
-		return errors.Wrap(err, "could not save config")
-	}
-
-	return nil
-}
-
-func (app *earthlyApp) printSatellites(satellites []cloud.SatelliteInstance, orgID string) {
-	for _, satellite := range satellites {
-		if satellite.Name == app.cfg.Satellite.Name && satellite.Org == orgID {
-			fmt.Printf("* %s\n", satellite.Name)
-		} else {
-			fmt.Printf("  %s\n", satellite.Name)
-		}
-	}
-}
-
-func (app *earthlyApp) getSatelliteOrgID(ctx context.Context, cloudClient cloud.Client) (string, error) {
-	// We are cheating here and forcing a re-auth before running any satellite commands.
-	// This is because there is an issue on the backend where the token might be outdated
-	// if a user was invited to an org recently after already logging-in.
-	// TODO Eventually we should be able to remove this cheat.
-	err := cloudClient.Authenticate(ctx)
-	if err != nil {
-		return "", errors.New("unable to authenticate")
-	}
-	var orgID string
-	if app.satelliteOrg == "" {
-		orgs, err := cloudClient.ListOrgs(ctx)
-		if err != nil {
-			return "", errors.Wrap(err, "failed finding org")
-		}
-		if len(orgs) == 0 {
-			return "", errors.New("not a member of any organizations - satellites only work within an org")
-		}
-		if len(orgs) > 1 {
-			return "", errors.New("more than one organizations available - please specify the name of the organization using `--org`")
-		}
-		app.satelliteOrg = orgs[0].Name
-		orgID = orgs[0].ID
-	} else {
-		var err error
-		orgID, err = cloudClient.GetOrgID(ctx, app.satelliteOrg)
-		if err != nil {
-			return "", errors.Wrap(err, "invalid org provided")
-		}
-	}
-	return orgID, nil
-}
-
-func (app *earthlyApp) actionSatelliteLaunch(cliCtx *cli.Context) error {
-	app.commandName = "launch"
-
-	if cliCtx.NArg() != 1 {
-		return errors.New("satellite name is required")
-	}
-
-	app.satelliteName = cliCtx.Args().Get(0)
-
-	cloudClient, err := cloud.NewClient(app.apiServer, app.sshAuthSock, app.authToken, app.console.Warnf)
-	if err != nil {
-		return errors.Wrap(err, "failed to create cloud client")
-	}
-
-	orgID, err := app.getSatelliteOrgID(cliCtx.Context, cloudClient)
-	if err != nil {
-		return err
-	}
-
-	app.console.Printf("Launching Satellite. This could take a moment...\n")
-	_, err = cloudClient.LaunchSatellite(cliCtx.Context, app.satelliteName, orgID)
-	if err != nil {
-		return errors.Wrapf(err, "failed to create satellite %s", app.satelliteName)
-	}
-	app.console.Printf("...Done\n")
-
-	err = app.useSatellite(cliCtx, app.satelliteName, app.satelliteOrg)
-	if err != nil {
-		return errors.Wrap(err, "could not configure satellite for use")
-	}
-	app.console.Printf("The satellite %s has been automatically selected for use. To go back to using local builds you can use\n\n\tearthly satellite unselect\n\n", app.satelliteName)
-
-	return nil
-}
-
-func (app *earthlyApp) actionSatelliteList(cliCtx *cli.Context) error {
-	app.commandName = "list"
-
-	if cliCtx.NArg() != 0 {
-		return errors.New("command does not accept any arguments")
-	}
-
-	cloudClient, err := cloud.NewClient(app.apiServer, app.sshAuthSock, app.authToken, app.console.Warnf)
-	if err != nil {
-		return errors.Wrap(err, "failed to create cloud client")
-	}
-
-	orgID, err := app.getSatelliteOrgID(cliCtx.Context, cloudClient)
-	if err != nil {
-		return err
-	}
-
-	satellites, err := cloudClient.ListSatellites(cliCtx.Context, orgID)
-	if err != nil {
-		return err
-	}
-
-	app.printSatellites(satellites, orgID)
-	return nil
-}
-
-func (app *earthlyApp) actionSatelliteDestroy(cliCtx *cli.Context) error {
-	app.commandName = "launch"
-
-	if cliCtx.NArg() != 1 {
-		return errors.New("satellite name is required")
-	}
-
-	app.satelliteName = cliCtx.Args().Get(0)
-
-	cloudClient, err := cloud.NewClient(app.apiServer, app.sshAuthSock, app.authToken, app.console.Warnf)
-	if err != nil {
-		return errors.Wrap(err, "failed to create cloud client")
-	}
-
-	orgID, err := app.getSatelliteOrgID(cliCtx.Context, cloudClient)
-	if err != nil {
-		return err
-	}
-
-	app.console.Printf("Destroying Satellite. This could take a moment...\n")
-	err = cloudClient.DeleteSatellite(cliCtx.Context, app.satelliteName, orgID)
-	if err != nil {
-		return errors.Wrapf(err, "failed to delete satellite %s", app.satelliteName)
-	}
-	app.console.Printf("...Done\n")
-
-	if app.satelliteName == app.cfg.Satellite.Name {
-		err = app.useSatellite(cliCtx, "", "")
-		if err != nil {
-			return errors.Wrapf(err, "failed unselecting satellite")
-		}
-		app.console.Printf("Satellite has also been unselected\n")
-	}
-	return nil
-}
-
-func (app *earthlyApp) actionSatelliteDescribe(cliCtx *cli.Context) error {
-	app.commandName = "describe"
-
-	if cliCtx.NArg() != 1 {
-		return errors.New("satellite name is required")
-	}
-
-	app.satelliteName = cliCtx.Args().Get(0)
-
-	cloudClient, err := cloud.NewClient(app.apiServer, app.sshAuthSock, app.authToken, app.console.Warnf)
-	if err != nil {
-		return errors.Wrap(err, "failed to create cloud client")
-	}
-
-	orgID, err := app.getSatelliteOrgID(cliCtx.Context, cloudClient)
-	if err != nil {
-		return err
-	}
-
-	satellite, err := cloudClient.GetSatellite(cliCtx.Context, app.satelliteName, orgID)
-	if err != nil {
-		return err
-	}
-
-	app.console.Printf("name: %s", satellite.Name)
-	app.console.Printf("version: %s", satellite.Version)
-	app.console.Printf("platform: %s", satellite.Platform)
-	app.console.Printf("status: %s", satellite.Status)
-	app.console.Printf("selected: %t", app.satelliteName == satellite.Name)
-	return nil
-}
-
-func (app *earthlyApp) actionSatelliteSelect(cliCtx *cli.Context) error {
-	app.commandName = "select"
-
-	if cliCtx.NArg() != 1 {
-		return errors.New("satellite name is required")
-	}
-
-	app.satelliteName = cliCtx.Args().Get(0)
-
-	cloudClient, err := cloud.NewClient(app.apiServer, app.sshAuthSock, app.authToken, app.console.Warnf)
-	if err != nil {
-		return errors.Wrap(err, "failed to create cloud client")
-	}
-
-	orgID, err := app.getSatelliteOrgID(cliCtx.Context, cloudClient)
-	if err != nil {
-		return err
-	}
-
-	satellites, err := cloudClient.ListSatellites(cliCtx.Context, orgID)
-	if err != nil {
-		return err
-	}
-
-	found := false
-	for _, s := range satellites {
-		if app.satelliteName == s.Name {
-			err = app.useSatellite(cliCtx, s.Name, app.satelliteOrg)
-			if err != nil {
-				return errors.Wrapf(err, "could not select satellite %s", app.satelliteName)
-			}
-			found = true
-			break
-		}
-	}
-
-	if !found {
-		return fmt.Errorf("%s is not a valid satellite", app.satelliteName)
-	}
-
-	app.printSatellites(satellites, orgID)
-	return nil
-}
-
-func (app *earthlyApp) actionSatelliteUnselect(cliCtx *cli.Context) error {
-	app.commandName = "unselect"
-
-	if cliCtx.NArg() != 0 {
-		return errors.New("command does not accept any arguments")
-	}
-
-	app.satelliteName = cliCtx.Args().Get(0)
-
-	err := app.useSatellite(cliCtx, "", "")
-	if err != nil {
-		return errors.Wrap(err, "could not unselect satellite")
-	}
-
-	return nil
-}
-
-func processSecrets(secrets, secretFiles []string, dotEnvMap map[string]string) (map[string][]byte, error) {
-	finalSecrets := make(map[string][]byte)
-	for k, v := range dotEnvMap {
-		finalSecrets[k] = []byte(v)
-	}
-	for _, secret := range secrets {
-		parts := strings.SplitN(secret, "=", 2)
-		key := parts[0]
-		var data []byte
-		if len(parts) == 2 {
-			// secret value passed as argument
-			data = []byte(parts[1])
-		} else {
-			// Not set. Use environment to fetch it.
-			value, found := os.LookupEnv(secret)
-			if !found {
-				return nil, errors.Errorf("env var %s not set", secret)
-			}
-			data = []byte(value)
-		}
-		if _, ok := finalSecrets[key]; ok {
-			return nil, errors.Errorf("secret %q already contains a value", key)
-		}
-		finalSecrets[key] = data
-	}
-	for _, secret := range secretFiles {
-		parts := strings.SplitN(secret, "=", 2)
-		if len(parts) != 2 {
-			return nil, errors.Errorf("unable to parse --secret-file argument: %q", secret)
-		}
-		k := parts[0]
-		path := fileutil.ExpandPath(parts[1])
-		data, err := os.ReadFile(path)
-		if err != nil {
-			return nil, errors.Wrapf(err, "failed to open %q", path)
-		}
-		if _, ok := finalSecrets[k]; ok {
-			return nil, errors.Errorf("secret %q already contains a value", k)
-		}
-		finalSecrets[k] = []byte(data)
-	}
-	return finalSecrets, nil
-}
-
-func defaultConfigPath() string {
-	earthlyDir := cliutil.GetEarthlyDir()
-	oldConfig := filepath.Join(earthlyDir, "config.yaml")
-	newConfig := filepath.Join(earthlyDir, "config.yml")
-	oldConfigExists, _ := fileutil.FileExists(oldConfig)
-	newConfigExists, _ := fileutil.FileExists(newConfig)
-	if oldConfigExists && !newConfigExists {
-		return oldConfig
-	}
-	return newConfig
 }
